### 2023-10-21 d34ab5a
Update paper.bib
Fixed invalid DOI
