+++ paper/paper.bib
@@ -126,7 +126,7 @@ number = {4},
 pages = {423-438},
 year = {1977},
 issn = {0022-4596},
-doi = {https://doi.org/10.1016/0022-4596(77)90020-2},
+doi = {10.1016/0022-4596(77)90020-2},
 url = {https://www.sciencedirect.com/science/article/pii/0022459677900202},
 author = {M. Ghedira and H. Vincent and M. Marezio and J.C. Launay}}
 
