### 2023-10-26 c2e7f89
Begun work on lightweight mode for the particle generation
