### 2023-10-27 d016e37
Implemented _lightweight_mode for the particle generator
