### 2023-10-30 b6c84b9
Updated DOIs in paper.bib
