+++ paper/paper.bib
@@ -46,7 +46,8 @@
   title={Underneath the Bragg peaks: structural analysis of complex materials},
   author={Egami, Takeshi and Billinge, Simon JL},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/c2010-0-66357-7}
 }
 
 @article{juhas2015complex,
@@ -57,7 +58,8 @@
   number={6},
   pages={562--568},
   year={2015},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
+  doi={10.1107/s2053273315014473}
 }
 
 @article{cervellino2010debussy,
@@ -94,14 +96,15 @@
 }
 
 @article{putnam2015bcl,
-  title={BCL:: SAXS: GPU accelerated D ebye method for computation of small angle X-ray scattering profiles},
+  title={BCL:: SAXS: GPU accelerated Debye method for computation of small angle X-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
   journal={Proteins: Struct., Funct., Genet.},
   volume={83},
   number={8},
   pages={1500--1512},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/prot.24838}
 }
 
 @article{Waasmaier:sh0059,
@@ -114,7 +117,6 @@ number = "3",
 pages = "416--431",
 month = "May",
 doi = {10.1107/S0108767394013292},
-url = {https://doi.org/10.1107/S0108767394013292},
 abstract = {New analytical X-ray scattering-factor representations valid for the full range of sin {$\Theta$}/{$\lambda$} from 0.0 to 6.0{\AA}${\sp {$-$}1}$ have been developed from fits of a linear combination of five Gaussians to the values of the scattering factors tabulated in {\it International Tables for Crystallography} (1992) [Vol. C. Dordrecht: Kluwer Academic Publishers]. The resulting functions for both neutral atoms and ions are compared with the existing parametrizations, which are applicable for 0.0{--}2.0{\AA}${\sp {$-$}1}$ and 2.0{--}6.0{\AA}${\sp {$-$}1}$, respectively. The quality of the new parametrization involving 11 parameters per atom (ion) compares well with the previous work or is even superior. Examples are discussed, some errors in {\it International Tables for Crystallography}, Vol. C are indicated and a warning is given that most of the previously published four Gaussian expansions for ions are inadequate for calculations involving sin {$\Theta$}/{$\lambda$} {$\ge$} 2.0{\AA}${\sp {$-$}1}$.},
 }
 
@@ -127,8 +129,8 @@ pages = {423-438},
 year = {1977},
 issn = {0022-4596},
 doi = {10.1016/0022-4596(77)90020-2},
-url = {https://www.sciencedirect.com/science/article/pii/0022459677900202},
-author = {M. Ghedira and H. Vincent and M. Marezio and J.C. Launay}}
+author = {M. Ghedira and H. Vincent and M. Marezio and J.C. Launay}
+}
 
 @misc{NVIDIA, url={https://resources.nvidia.com/en-us-grace-cpu/grace-hopper-superchip}, journal={NVIDIA}, author={NVIDIA}, year={2023}} 
 
