### 2023-10-30 86020ef
Added high-level implementation details
