+++ requirements.txt
@@ -6,3 +6,4 @@ torchvision==0.15.2
 PyYAML==6.0.1
 ipywidgets==8.1.1
 tqdm==4.66.1
+distutils
