+++ requirements.txt
@@ -1,3 +1,4 @@
+distutils
 ase==3.22.1
 numpy==1.24.2
 torch==2.0.1
@@ -6,4 +7,3 @@ torchvision==0.15.2
 PyYAML==6.0.1
 ipywidgets==8.1.1
 tqdm==4.66.1
-distutils
