+++ requirements.txt
@@ -1,4 +1,3 @@
-distutils
 ase==3.22.1
 numpy==1.24.2
 torch==2.0.1
