### 2023-10-30 #5964 @FrederikLizakJohansen
Hi @phibeck 👋, thank you for your feedback.

- As you can see, we have addressed the missing and invalid DOIs. They are no longer flagged by @editorialbot.
- Thank you for your suggestion on moving some of the function documentation into the repository - we have done exactly that.
- We see that you have already found some reviewers. We have also identified some relevant reviewers, three of which are contributors to the Mantid project, which share a common focus on supporting the processing and analysis of materials science data, albeit with different specific functionalities and target areas. All five potential reviewers have expertise in 'scattering' which we think are very relevant to review the DebyeCalculator software and paper.
-- **stuartcampbell** (Contributor to Mantid, relevant topics: Neutron scattering, X-ray techniques)
-- **mdoucet** (Contributor to Mantid, relevant topics: Physics, Neutron Scattering, Materials Science)
-- **jwuttke** (Contributor to Mantid, relevant topics: Scattering, Diffraction)
-- **marcocamma** (Employed at the European Synchrotron Radiation Facility, relevant topics: X-ray, Large facilities, Scattering)
-- **KedoKedo** (relevant topics: Neutron diffraction, X-ray diffraction, Data visualization)
