+++ setup.py
@@ -9,20 +9,16 @@ reqs = [str(ir.req) for ir in install_reqs]
 
 setup(
     name='DebyeCalculator',
-    version='0.1',
-    description='A fast vectorized implementation of the Debye equation on both CPU and GPU.',
+    python_requires='>=3.9,<3.11'
+    version='1.0.1',
+    description='A vectorised implementation of the Debye Equation on CPU and GPU',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
     author='Frederik L. Johansen and Andy S. Anker',
     author_email='frjo@di.ku.dk and andy@chem.ku.dk',
     install_requires=reqs,
     packages=find_packages(
-        where='DebyeCalculator',
-        exclude=[
-            'generate_nanoparticles*', 
-            'produce_figure_data*', 
-            'SASCalculator*'
-        ],
+        where='debyecalculator',
     ),
     include_package_data=True,
     package_data={'': ['*.yaml']},
@@ -30,6 +26,6 @@ setup(
     license='Apache 2.0',
     license_files=['LICENSE*'],
     classifiers=[
-        'Programming Language :: Python'
+        'Programming Language :: Python :: 3'
     ]
 )
