### 2023-10-31 1d50d2b
Added _lightweight_mode to main class
