### 2023-10-31 #5964 @phibeck
Hi @FrederikLizakJohansen,
>     * As you can see, we have addressed the missing and invalid DOIs. They are no longer flagged by @editorialbot.
>     * Thank you for your suggestion on moving some of the function documentation into the repository - we have done exactly that.
Great, looks good. Thanks for the reviewer suggestions. I will come back to this depending on the response to my first invitation.
