### 2023-10-31 #5964 @marcocamma
Dear @phibeck and all,
The topic looks very interesting ! I have written also some simple code for Debye scattering calculation ... that was not particularly fast ... so I am looking forward to the GPU version !
I will be happy to review the manuscript,
Looking forward to further instructions.
