### 2023-11-01 53d6799
Removed unneccessary instances of _lightweight_mode arguments
