### 2023-11-06 #6024 @editorialbot
[REVIEW]: A GPU-Accelerated Open-Source Python Package for Calculating Powder Diffraction, Small-Angle-, and Total Scattering with the Debye Scattering Equation
**Submitting author:** @AndySAnker (<a href="http://orcid.org/0000-0002-7403-6642">Andy Sode Anker</a>)
**Repository:** https://github.com/FrederikLizakJohansen/DebyeCalculator
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.7
