### 2023-11-17 421f11f
Narrowed margin on floating atoms to avoid special cases of floating atoms
