### 2023-11-28 7058aae
Begun work to make NP cutting more general
