### 2023-11-28 FrederikLizakJohansen/DebyeCalculator#13 @KedoKudo
[JOSS review] Add tags and release on Github
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

Currently there is no version tag or software releases available on Github, but the installed package does not have version.
It would be great if the developers can:

- use existing toolkits like `versioninggit` to automatically track versions
- add version to the published package
- tag a release on GitHub
