### 2023-11-28 FrederikLizakJohansen/DebyeCalculator#14 @KedoKudo
[JOSS review] Add documentation and publish it
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

Currently the software does not have a documentation. The usage and instructions are in the readme file and there is no API reference doc.

It would be great if the developers can:
- add a documentation folder like `docs`
- use sphinx or other software to help generate the API doc
- publish the doc on readthedoc or on GitHub page.

Also, it would be great to provide a step-by-step guild on how to use the software in the documentation, along with necessary screenshots.
