### 2023-11-28 FrederikLizakJohansen/DebyeCalculator#15 @KedoKudo
[JOSS review] GoogleColab example does not show plots
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

After clicking the link to Google collab and running through the two cells. The widget shows up, but the generate plot button does not trigger the simulation and generate corresponding plots.
