### 2023-11-28 FrederikLizakJohansen/DebyeCalculator#16 @KedoKudo
[JOSS review] benchmark on Apple silicon
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

It would be interesting to see how the software performs on Apple silicon when running as a CPU process and using the `mps` backend.
