### 2023-11-28 #6024 @KedoKudo
I created a few issues after the first round of review:

- https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/16
- https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/15
- https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/14
- https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/13

I will continue the review after these issues are addressed 😃
