### 2023-11-28 FrederikLizakJohansen/DebyeCalculator#17 @marcocamma
Improving installation instructions
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

I would improve the installation instructions as follow:
- install python version of interest (if needed)
  - create virtual environment: ```python3.11 -m venv debyecalculator_venv```
  - activate it ```source ./debyecalculator_venv/bin/activate``` (with bash)
  - install with pip from
      - the Python Package Index : ```pip install DebyeCalculator```
      - github ```pip install git+https://github.com/FrederikLizakJohansen/DebyeCalculator.git```
