### 2023-11-29 #6024 @marcocamma
So I spent a bit of time working with the debyecalculator.
- I have to say that the installation has been less straightforward that I would have hoped:
  - in none of the computer I has the right python version (so I had to install it ... that needs some time)
  - in those computers with GPU that I have access to, the NVIDIA driver was too old so I could not really test GPU implementation
  - can the author clairfy why such a tight constrain on the python version ?
  - it makes me worry that unless the package is very well maintained, it may quickly become difficult to install/use
Here my comments on the library iteself:
- I would improve a bit the bench-marking with a script that test
  - with/without GPU
  - structures with different sizes
  - test different batch_size
- execution times on a decent device could be added to the script to have some reference value.
- this script would quickly allow to test how fast the program can run on a particular device... or even test different devices to see where it might make more sense to run long calculations.
-  I am less and less fun of simple tuples as return values ... did the authors consider a named_tuple (or a dictionary instead) ?
- last, but not least, I find that providing the input structure as CIF or XYZ file is a limitation
  - I would add the possibility of providing the input as tuple (atoms_list,xyz_array)
    - If one has thousands of structures (for example from MD simulations) that are neatly stored in a file, it would be a pity to write one file per structure just to calculate the G(r) or S(q)
  - related to the point above, would it make sense to move generate_nanoparticles as standalone utility function.
    - it could take a cif and a radius and return atom_list,xyz_position ?
- I think it would be nice to be able to calculate the partial g(r). Of course this can be done by "filtering the inputs" (one more reasons to have non-file based atom_list,xyz_positions). But maybe g(r) could return the partial g(r) as well ?

As for the paper itself:
- I think it would make sense to spend a bit more words on the comparison with other packages in terms of capabilities.
- Also benchmark comparison with other software (and not only with DiffPy-CMI) would be welcomes.
- Different softwares/papers use different definitions for G(r), S(q), etc. It is important to add what is used for this package

I wait for changes/comment before doing a second round. Keep up the good work !
