+++ README.md
@@ -111,7 +111,7 @@ __Andy S. Anker__<sup>1</sup>
 <sup>2</sup> Department of Computer Science, University of Copenhagen, Denmark
 
 Should there be any questions, desired improvements or bugs please contact us on GitHub or 
-through email: __frjo@di.ku.dk__ and __andy@chem.ku.dk__.
+through email: __frjo@di.ku.dk__ and __ansoan@dtu.dk__.
 
 # Cite
 If you use our code or our results, please consider citing our paper. Thanks in advance!
