### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#18 @AndySAnker
[JOSS review] Installation: Allow additional Python version
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

@marcocamma suggests allowing DebyeCalculator to be installed on additional Python versions.
