### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#19 @AndySAnker
[JOSS review] Bench-marking script
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

@marcocamma suggests to include a bench-marking script which tests:
- with/without GPU
- structures with different sizes
- test different batch_size
- execution times on a decent device could be added to the script to have some reference value.

this script would quickly allow to test how fast the program can run on a particular device... or even test different devices to see where it might make more sense to run long calculations.
