### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#20 @AndySAnker
[JOSS review] Named_tuple
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

@marcocamma suggests using a named_tuple or dictionary as return values instead of simple tuples.
