### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#21 @AndySAnker
[JOSS review] Various formats for input structure
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

@marcocamma suggests to add functionality of providing the input as tuple (atoms_list,xyz_array).

related to the point above, would it make sense to move generate_nanoparticles as standalone utility function.
it could take a cif and a radius and return atom_list,xyz_position ?
