### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#22 @AndySAnker
[JOSS review] Partial Gr
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

@marcocamma suggests adding a function that can calculate the partial Gr.
