### 2023-12-03 922d658
Update paper.bib
Added DISCUS references
