+++ paper/paper.bib
@@ -95,6 +95,28 @@
   publisher={International Union of Crystallography}
 }
 
+@article{proffen1999discus,
+  title={DISCUS, a program for diffuse scattering and defect structure simulations--update},
+  author={Proffen, Th and Neder, RB},
+  journal={Journal of Applied Crystallography},
+  volume={32},
+  number={4},
+  pages={838--839},
+  year={1999},
+  publisher={International Union of Crystallography}
+}
+
+@article{proffen1997discus,
+  title={DISCUS: A program for diffuse scattering and defect-structure simulation},
+  author={Proffen, Thomas and Neder, Reinhard B},
+  journal={Journal of applied crystallography},
+  volume={30},
+  number={2},
+  pages={171--175},
+  year={1997},
+  publisher={International Union of Crystallography}
+}
+
 @article{putnam2015bcl,
   title={BCL:: SAXS: GPU accelerated Debye method for computation of small angle X-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
