### 2023-12-03 b309421
Update paper.md
Changes to Statement of Need section:
- Changed TOPAS to DISCUS.
- Added the importance of DebyeCalculator is written in Python.
