+++ paper/paper.md
@@ -90,8 +90,8 @@ The choice of GPU hardware has a substantial influence on this speed advantage. 
 
 # Statement of need
 
-Several software packages already exist for simulating the Debye scattering equation, including DiffPy-CMI[@juhas2015complex], debyer[@debyer], Debussy[@cervellino2010debussy; @cervellino2015debussy], TOPAS[@coelho2018topas], and BCL::SAXS[@putnam2015bcl]. Our software distinguishes itself in several ways. Firstly, it is freely available and open-source licensed under the Apache License 2.0. Moreover, it is conveniently implemented as a ‘pip’ install package and has been integrated with Google Colab[https://github.com/FrederikLizakJohansen/DebyeCalculatorGPU/blob/main/quickstart/QuickStart.ipynb], allowing users to rapidly calculate PD, TS, SAS, and PDF data using the Debye scattering equation without the need of local software installations. ´DebyeCalculator´ can be run through an interactive interface (see \autoref{fig:figure_2}), where users can calculate $I(Q)$, $S(Q)$, $F(Q)$, and $G(r)$ from structural models on both CPU and GPU.
-
+Several software packages already exist for simulating the Debye scattering equation, including DiffPy-CMI[@juhas2015complex], debyer[@debyer], Debussy[@cervellino2010debussy; @cervellino2015debussy], DISCUS[@proffen1997discus, proffen1999discus], and BCL::SAXS[@putnam2015bcl]. 'DebyeCalculator' stands out as it is open-source, Python-based, and GPU-compatible. Python's status as an extremely popular and accessible language in scientific computing makes it straightforward for researchers to quickly perform Debye scattering calculations while ensuring easy integration with other data science tools.
+'DebyeCalculator' is open-source licensed under the Apache License 2.0. Moreover, it is conveniently implemented as a ‘pip’ install package and has been integrated with Google Colab[https://github.com/FrederikLizakJohansen/DebyeCalculatorGPU/blob/main/quickstart/QuickStart.ipynb], allowing users to rapidly calculate PD, TS, SAS, and PDF data using the Debye scattering equation without the need of local software installations. ´DebyeCalculator´ can also be run through an interactive interface (see \autoref{fig:figure_2}), where users can calculate $I(Q)$, $S(Q)$, $F(Q)$, and $G(r)$ from structural models on both CPU and GPU. 
 ![The interact mode of ´DebyeCalculator´ provides a one-click interface, where the user can update parameters and visualise $I(Q)$, $S(Q)$, $F(Q)$, and $G(r)$. Additionally, the $I(Q)$, $S(Q)$, $F(Q)$, $G(r)$, and xyz-file can be downloaded, including metadata.\label{fig:figure_2}](../figures/figure_2.png)
 
 # Acknowledgements
