### 2023-12-03 123935a
Update paper.md
fixed typos
