### 2023-12-03 ddfe565
Moved implementation details into docs folder
