### 2023-12-03 39553ce
Updated docs/Equations.md and High_level_implementation_details.md
