### 2023-12-03 2c0896b
Updated docs/Equations.md
