### 2023-12-03 2ab33dd
Updated Equations.md
