### 2023-12-03 c9630b0
Updated Equations.md
