+++ docs/Equations.md
@@ -11,14 +11,17 @@ In this equation $Q$ is the momentum transfer of the scattered radiation, $N$ is
 # The Total Scattering Structure Function - $S(Q)$
 
 The Total Scattering Structure Functionm, $S(Q)$, is calculated as:
+
 $$
 S(Q) = \frac{I_{\text{coh}}(Q) + \langle b(Q) \rangle^2 - \langle b(Q)^2 \rangle}{N \langle b(Q) \rangle^2} - 1
 $$
+
 Where $I_{coh}$ is the scattering intensity as we only simulate the coherent scattering signal.
 
 # The Reduced Total Scattering Function - $F(Q)$
 
 The Reduced Total Scattering Function, $F(Q)$, is calculated as:
+
 $$
 F(Q) = Q \left( S(Q) \right)
 $$
@@ -26,7 +29,9 @@ $$
 # The Reduced Atomic Pair Distribution Function - $G(r)$
 
 The Reduced Atomic Pair Distribution Function, $G(r)$, is calculated as:
+
 $$
 G(r) = \frac{2}{\pi} \int_{Q_{\text{min}}}^{Q_{\text{max}}} F(Q) \sin(Qr) \dQ
 $$
+
 Where $Q_{min}$ and $Q_{max}$ is the minimum and maximum Q-values of the data.
