### 2023-12-03 c8bbf3f
Updated Equations.md
