+++ README.md
@@ -23,8 +23,10 @@ Here, we provide an optimised code for the calculation of the Debye scattering e
 # Installation
 
 `DebyeCalculator` requires Python version >=3.9, <3.12. If needed, create an environment with any of these Python versions:
+
 ```
-conda create -n debyecalculator_env python=3.9```
+conda create -n debyecalculator_env python=3.9
+```
 
 ```
 conda activate debyecalculator_env
