### 2023-12-03 FrederikLizakJohansen/DebyeCalculator#17 @AndySAnker
Thank you for the great suggestion!

We have updated the README.md file to include the possibility of creating an environment with the proper Python version before pip install. I close this issue - let us know if it should be reopened.
