### 2023-12-04 #6024 @AndySAnker
Thanks for the great feedback @marcocamma and @KedoKudo. We will work through the points one by one during the next weeks. Thanks for organising the suggestions as issues that makes it easier for us to organise responses.

Response to @KedoKudo will happen in the respective issues.

Response to @marcocamma follows here and at the respective issues which we have generated from your comments.

### Technical comments from @marcocamma
**Python version:**
We have moved this to an issue here: https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/18

**Bench-marking script:**
We have moved this to an issue here: https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/19
You write "execution times on a decent device could be added to the script to have some reference value.". Can you clarify what you mean with "a decent device"? We perform execution times of three different devices (x86-64 CPU with 6GB, NVIDIA RTX A3000 Laptop GPU with 6GB and NVIDIA Titan RTX GPU with 24GB). We consider all these three devices as decent.

**Did the authors consider a named_tuple (or a dictionary instead) as return value?**
We have moved this to an issue here: https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/20

**I would add the possibility of providing the input as tuple (atoms_list,xyz_array)**
We have moved this to an issue here: https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/21

**Adding the calculation of partial g(r) to the package**
We have moved this to an issue here: https://github.com/FrederikLizakJohansen/DebyeCalculator/issues/22

### As for the paper itself:
**I think it would make sense to spend a bit more words on the comparison with other packages in terms of capabilities.
Also benchmark comparison with other software (and not only with DiffPy-CMI) would be welcomes.**
The DebyeCalculator is designed not merely to be the fastest tool for calculating scattering patterns using the Debye scattering equation. Instead, our goal is to develop a multifaceted software with several key features:
1) **Python-based**. Python has become extremely popular and therefore this ensures that it is easier for newcomers to use and also that it can be integrated into other data science tools.
2) **Easy to use** to calculate both powder diffraction, total scattering with pair distribution function data and small-angle scattering data. In fact, we have already been contacted by a researcher, who uses DebyeCalculator for educational purposes as the interact mode allows students to easily grasp the impact of various parameters.
3) **GPU-compatible**. With the growing interest in using Machine Learning (ML) for scattering data analysis (https://pubs.rsc.org/en/content/articlelanding/2023/SC/D3SC05081E), GPU compatibility becomes crucial. While current software can calculate scattering patterns for ML training, they lack the ability to backpropagate through these calculations. DebyeCalculator addresses this gap, being both Python-based and GPU-compatible. Note that the GPU calculations are handled in PyTorch, which is a highly popular ML library.
4) **Fast**. While speed is not our primary focus, it remains an important aspect. We specifically benchmark DebyeCalculator against DiffPy-CMI because DiffPy-CMI is Python-based. Debyer is also Python-based, but it is not maintained anymore: https://github.com/wojdyr/debyer/issues/2. BCL::SAXS is intended for small-angle scattering data. Debussy and DISCUS, running through GUI's, do not (to our knowledge) offer Python script compatibility.

**Different softwares/papers use different definitions for G(r), S(q), etc. It is important to add what is used for this package**
Thank you for the suggestion, we have added a docs/ folder which now also contains this documentation.
