### 2023-12-04 7447d29
Update Equations.md
