### 2023-12-04 ae59332
Update Equations.md
