+++ docs/Equations.md
@@ -13,7 +13,7 @@ In this equation $Q$ is the momentum transfer of the scattered radiation, $N$ is
 The Total Scattering Structure Functionm, $S(Q)$, is calculated as:
 
 $$
-S(Q) = \frac{1}{N} \frac{I(Q)}{\langle b(Q) \rangle^2}
+S(Q) = \frac{1}{N} \frac{I(Q)}{\langle b \rangle^2}
 $$
 
 The normalisation is performed over the compositional average of the form factor (scattering length) contributions, $\langle b \rangle^2 = \sum_µ c_µ b_µ$, where $c_µ$ is the contribution fraction of element $µ$ as well as the number of atoms $N$. Certain conventions will have the total scattering structure function converge to unity at large $Q$. Our implementation is different in that $S(Q)$ will converge to nullity. This however can be rectified by simply adding unity to the above equation.
