### 2023-12-04 e47ce2b
Update Equations.md
