### 2023-12-04 8f93836
Update Equations.md
