### 2023-12-04 f23a942
Begun implementing new general NP cutting
