### 2023-12-07 1eafdb5
Implemented more general NP cutting. Awaiting final testing of functionality
