### 2023-12-08 68d5d9a
Implemented and tested more general NP cutting. Can now also returns edges and distances if needed.
