### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#24 @elena-pascal
[JOSS review] Instructions missing for how the software wants to access the GPU
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

I followed the install instructions which ran cleanly. I have an NVIDIA RTX card on this PC. I expected then that I can run
```
calc = DebyeCalculator(device='cuda')
```
actually, I expected to be able to run `device='gpu'` since the class documentation only talks about cpu/gpu.

> (e.g., 'cuda' for GPU or 'cpu' for CPU). Default is 'cuda' if the computer has a GPU.

But I actually don't have `cuda` and I get ```AssertionError("Torch not compiled with CUDA enabled")```

It would be nice to have either  more detailed requirements, specifically on GPU drivers and software needed or more details instructions on how to achieve the GPU interface layers DebyeCalculator expects.

I think either send the user to install `torch` following the [website instructions,](https://pytorch.org/get-started/locally/) or offer the instructions for the platform you want to support. `pip install torch` is not guaranteed to give you something that can talk to the GPU on any platform configuration under the sun.
