### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#25 @elena-pascal
[JOSS review] Instructions missing on running tests
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

This is an easy one, but adding the line needed to run the tests in the instructions would be nice.
