### 2023-12-12 3591003
Support for python 3.7 & 3.8 based on #18
