### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#18 @FrederikLizakJohansen
Changed .toml file to support python 3.7 and 3.8 in 3591003
