### 2023-12-12 919cca3
fixed small issue with lightweight mode
