+++ .github/workflows/unit-tests.yml
@@ -2,16 +2,20 @@ name: Unit tests
 
 on:
   push:
-    branches: [ main ]
+    branches:
+      - main
+      - poetry_unit_tests
   pull_request:
-    branches: [ main ]
+    branches: 
+      - main
+      - poetry_unit_tests
 
 jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v2
@@ -19,11 +23,15 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
+    - name: Setup poetry and pytest
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt
+        pip install poetry
         pip install pytest
+    - name: Install dependencies
+      run: |
+        poetry env use python
+        poetry update
     - name: Run tests
       run: |
-        pytest 
+        poetry pytest
