+++ .github/workflows/unit-tests.yml
@@ -28,8 +28,10 @@ jobs:
         python -m pip install --upgrade pip
         pip install poetry
         pip install pytest
-    - name: Install dependencies and run
+    - name: Install dependencies
       run: |
         poetry env use python
         poetry update
+    - name: Run pytest
+      run: |
         poetry run pytest
