### 2023-12-12 7f30b98
FileNotFound error handled better
