### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#18 @FrederikLizakJohansen
- Unit tests now uses Poetry dependency management akin to the PyPI installation for better consistency, see c258055
- Pytest added as dev-dependency, see 814894c
