### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#18 @FrederikLizakJohansen
Updated PyPI project to version 1.0.2 (https://pypi.org/project/debyecalculator/) that now supports python 3.7 and python 3.8 as well as the previously supported versions
