### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#18 @FrederikLizakJohansen
Thank you for a great suggestion! 🚀
Closing this issue. If you encounter any further problems, feel free to reopen or create a new issue.
