### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#26 @elena-pascal
[JOSS review] Too quiet defaulting to CPU
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

Given that the novelty in this package is the GPU performance, I think it is a shame that it doesn't let the user know through any level of warning that they are running the computation on the CPU and not benefiting from the GPU acceleration.

I'd consider throwing a warning for `torch.cuda.is_available()` returning False.
