### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#27 @elena-pascal
[JOSS review] Names could be less ambiguous
This is part of the review feedback for JOSS submission (https://github.com/openjournals/joss-reviews/issues/6024)

I am going to keep track here of user facing variables that could be more precise.

- [ ] `structure_path` is not just the path but the file name as well
I think naming it `structure_file` would be more clear. If you do want to emphasise that a path can be provided it seem that you have the code for that in `DebyeCalculator().interact()`

All the computation return tuples of parameters in reverse order from the name of the method. Is this on purpose?
- [ ] `calc.iq` returns Q, I
- [ ] `calc.sq` returns Q, S
- [ ] `calc.fq` returns Q, F
- [ ] `calc.gr` returns r, G
Another reviewer mentioned that `namedtuples` would improve this situation and I agree, I think giving outputs more precise names,  together with `namedtuples` for the two parameters would make the output more intuitive; ie. `calc.scattering_intensity()` would return `ScatteringIntensity(values, points)/ (y, x)/(ordinate, abscissa)...`. It also feels like those second return parameters are optional (?)
