### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#28 @elena-pascal
[JOSS review] API a bit cluncky
This is part of the review feedback for JOSS submission (openjournals/joss-reviews#6024)

From the usage example and available documentation it is not clear what is the benefit of passing the file at the method level and not at the object level together with the rest of the variables. It feel a bit clunky to pass the actual data after all the computation parameters have been initialised and computed.
