### 2023-12-12 FrederikLizakJohansen/DebyeCalculator#29 @elena-pascal
[JOSS review] Access to data used for the paper figure
This is part of the review feedback for JOSS submission (openjournals/joss-reviews#6024)

The paper could link to the exact data used for the figure or describe how data was generated. If the latter, it would still be nice to have access to a dataset that nicely highlights the gain in speed when moving to the GPU.
