### 2023-12-13 6a245d4
Better handling of FileNotFoundError
