### 2023-12-13 9ada2fc
Added absolute tolerance
