+++ README.md
@@ -22,7 +22,7 @@ Here, we provide an optimised code for the calculation of the Debye scattering e
 
 # Installation
 
-`DebyeCalculator` requires Python version >=3.9, <3.12. If needed, create an environment with any of these Python versions:
+`DebyeCalculator` requires Python version >=3.7, <3.12. If needed, create an environment with any of these Python versions:
 
 ```
 conda create -n debyecalculator_env python=3.9
@@ -34,14 +34,19 @@ conda activate debyecalculator_env
 
 ## Install with [pip](https://pypi.org/project/DebyeCalculator/)
 
-Run the following command to install the __DebyeCalculator__ package. (**Requires**: Python >=3.9, <3.12)
+Run the following command to install the __DebyeCalculator__ package. (**Requires**: Python >=3.7, <3.12)
 ```
-pip install DebyeCalculator
+pip install debyecalculator
 ```
 
 ## Install locally
 
-Run the following command to install the __DebyeCalculator__ package. (**Requires**: Python >=3.9, <3.12)
+Clone the repo
+```
+git clone https://github.com/FrederikLizakJohansen/DebyeCalculator.git
+```
+
+Run the following command to install the __DebyeCalculator__ package. (**Requires**: Python >=3.7, <3.12)
 ```
 pip install .
 or
