### 2023-12-13 0900726
Requirements and dependencies moved to pyproject.toml file
