### 2023-12-13 8beadc6
Added section about testing the installation, based on #25
