### 2023-12-13 FrederikLizakJohansen/DebyeCalculator#25 @FrederikLizakJohansen
Thank you for the suggestion!
The change was added in 8beadc6

I am closing this issue here. Feel free to reopen the issue if you would like to see any further changes made to this.
