### 2023-12-13 2cccef4
Moved parameter assertions to seperate function, better test coverage
