### 2023-12-13 d34cbfc
Restructured iq() function, introduced new namedtuple-function-structure, work-in-progress
