+++ debyecalculator/debye_calculator.py
@@ -6,7 +6,8 @@ import pkg_resources
 import warnings
 from glob import glob
 from datetime import datetime
-from typing import Union, Tuple, Any, List
+from typing import Union, Tuple, Any, List, Type
+from collections import namedtuple
 
 import torch
 from torch import cdist
@@ -29,6 +30,23 @@ from tqdm.auto import tqdm
 
 import collections
 import timeit
+
+StructureTuple = namedtuple('StructureTuple', 'elements size occupancy xyz triu_indices unique_inverse unique_form_factors form_avg_sq structure_inverse')
+IQTuple = namedtuple('IQTuple', 'Q I')
+
+ArrayLike = Union[np.ndarray, torch.Tensor]
+IntArrayLike = Union[List[int], np.ndarray, torch.Tensor]
+StructureSourceType = Union[
+    Tuple[List[str], ArrayLike],
+    Tuple[IntArrayLike, ArrayLike],
+    List[Tuple[List[str], ArrayLike]],
+    List[Tuple[IntArrayLike, ArrayLike]],
+    str,
+    List[str],
+    Atoms,
+    List[Atoms],
+]
+
 
 class DebyeCalculator:
     """
@@ -189,127 +207,131 @@ class DebyeCalculator:
             self.q = torch.arange(self.qmin, self.qmax, self.qstep).unsqueeze(-1).to(device=self.device)
             self.r = torch.arange(self.rmin, self.rmax, self.rstep).unsqueeze(-1).to(device=self.device)
 
-    def _initialise_structures(
+    def _initialise_structure(
         self,
-        structure_path: Union[str, Atoms, List[Atoms]],
+        structure_source: StructureSourceType,
         radii: Union[List[float], float, None] = None,
         disable_pbar: bool = False,
     ) -> None:
 
         """
-        Initialise atomic structures and unique elements form factors from an input file.
+        Initialise a single atomic structure and unique elements form factors from an input file.
 
         Parameters:
-            structure_path (Union[str, Atoms, List[Atoms]]): Path to the atomic structure file in XYZ/CIF format or stored ASE Atoms objects.
+            structure_source (StructureSourceType): Atomic structure source in XYZ/CIF format, ASE Atoms object or as a tuple of (atomic_identities, atomic_positions)
             radii (Union[List[float], float, None]): List/float of radii/radius of particle(s) to generate with parsed CIF
         """
-        # Check if input is a file or ASE Atoms object
-        if isinstance(structure_path, str):
-            # Check file and extention
-            structure_ext = structure_path.split('.')[-1]
-            if structure_ext not in ['xyz', 'cif']:
-                raise TypeError('FAILED: Invalid file/file-extention, accepts only .xyz or .cif data files')
-        elif isinstance(structure_path, Atoms) or all(isinstance(lst_elm, Atoms) for lst_elm in structure_path):
-            structure_ext = 'ase'
-        else:
-            raise TypeError('FAILED: Invalid structure format, accepts only .xyz, .cif data files or ASE Atoms objects')
 
-        # If cif, check for radii and generate particles
-        if structure_ext == 'cif':
-            if radii is not None:
-                ase_structures, _ = self.generate_nanoparticles(structure_path, radii, disable_pbar=disable_pbar, _lightweight_mode=self._lightweight_mode)
-                self.num_structures = len(ase_structures)
-            else:
-                raise ValueError('FAILED: When providing .cif data file, please provide radii (Union[List[float], float]) to generate from.')
-
-            self.struc_elements = []
-            self.struc_size = []
-            self.struc_occupancy = []
-            self.struc_xyz = []
[…]
