### 2023-12-14 620c6b8
Major restructuring of all scattering functions, handling of input and type-checking
