+++ debyecalculator/debye_calculator.py
@@ -32,7 +32,11 @@ import collections
 import timeit
 
 StructureTuple = namedtuple('StructureTuple', 'elements size occupancy xyz triu_indices unique_inverse unique_form_factors form_avg_sq structure_inverse')
-IQTuple = namedtuple('IQTuple', 'Q I')
+IqTuple = namedtuple('IqTuple', 'q i')
+SqTuple = namedtuple('SqTuple', 'q s')
+FqTuple = namedtuple('FqTuple', 'q f')
+GrTuple = namedtuple('GrTuple', 'r g')
+AllTuple = namedtuple('AllTuple', 'r q i s f g')
 
 ArrayLike = Union[np.ndarray, torch.Tensor]
 IntArrayLike = Union[List[int], np.ndarray, torch.Tensor]
@@ -46,7 +50,6 @@ StructureSourceType = Union[
     Atoms,
     List[Atoms],
 ]
-
 
 class DebyeCalculator:
     """
@@ -247,8 +250,8 @@ class DebyeCalculator:
             # Calculate average squared form factor and self scattering inverse indices
             counts = torch.from_numpy(counts).to(device=self.device)
             compositional_fractions = counts / torch.sum(counts)
-            struc_form_avg_sq = torch.sum(compositional_fractions.reshape(-1,1) * unique_form_factors, dim=0)**2
-            struc_inverse = torch.from_numpy(np.array([inverse[i] for i in range(size)])).to(device=self.device)
+            form_avg_sq = torch.sum(compositional_fractions.reshape(-1,1) * unique_form_factors, dim=0)**2
+            structure_inverse = torch.from_numpy(np.array([inverse[i] for i in range(size)])).to(device=self.device)
 
             return triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse
 
@@ -283,7 +286,7 @@ class DebyeCalculator:
                         occupancy = torch.from_numpy(structure[:,-1]).to(device=self.device, dtype=torch.float32)
                         xyz = torch.tensor(structure[:,1:-1].astype('float')).to(device=self.device, dtype=torch.float32)
                     else:
-                        occupancy = torch.ones((structure_size[0]), dtype=torch.float32).to(device=self.device)
+                        occupancy = torch.ones((size), dtype=torch.float32).to(device=self.device)
                         xyz = torch.tensor(structure[:,1:].astype('float')).to(device=self.device, dtype=torch.float32)
                 except:
                     raise IOError(f'Encountered invalid file format when trying to load structure from {structure_source}')
@@ -335,21 +338,17 @@ class DebyeCalculator:
         radii: Union[List[float], float, None] = None,
         keep_on_device: bool = False,
         _total_scattering: bool = False,
-    ) -> Union[Tuple[np.float32, Union[List[np.float32], np.float32]], Tuple[torch.FloatTensor, Union[List[torch.FloatTensor], torch.FloatTensor]]]:
-
+    ) -> Union[IqTuple, List[IqTuple]]:
         """
-        Calculate the scattering intensity I(Q) for the given atomic structure.
+        Calculate the scattering intensity I(Q) for the given atomic structure(s)
 
         Parameters:
-            structure_path (Union[str, Atoms, List[Atoms]]): Path to the atomic structure file in XYZ/CIF format or stored ASE Atoms objects.
-            radii (Union[List[float], float, None]): List/float of radii/radius of particle(s) to generate with parsed CIF
-            keep_on_device (bool): Flag to keep the results on the class device. Default is False.
-            _total_scattering (bool): Flag to return the scattering intensity I(Q) without the self-scattering contribution. Default is False.
+            structure_source (StructureSourceType): Atomic structure source in XYZ/CIF format, ASE Atoms object or as a tuple of (atomic_identities, atomic_positions)
[…]
