### 2023-12-14 FrederikLizakJohansen/DebyeCalculator#21 @FrederikLizakJohansen
Thank you for the suggestion. We think this is very relevant to how the package will be used in the future, which is why I am currently doing some major restructuring of the way that we handle input and output of the five functions: `iq`, `sq`, `fq`, `gr` and `_get_all` (on branch [issue-21](https://github.com/FrederikLizakJohansen/DebyeCalculator/tree/issue-21)), including the possibility of providing input as tuples (atoms_list,xyz_array). The restructuring naturally incorporates the suggestion (#20) of using namedtuples to better organize the various outputs of each function. Most recent commit can be found here 620c6b8

An update regarding `generate_nanoparticles` will follow
