+++ debyecalculator/form_factor_coef.yaml
@@ -1,231 +1,0 @@
-#  The atomic scattering factor is calculated using the method developed by  D. Waasmaier & A. Kirfel 
-#                                                                        
-#          New Analytical Scattering Factor Functions for Free Atoms       
-#                    and Ions for Free Atoms and Ions                      
-#                       D. Waasmaier & A. Kirfel                           
-#                    Acta Cryst. (1995). A51, 416-431                      
-#                                                                          
-#  fo the non-dispersive part of the atomic scattering factor is a           
-#  function of the selected element and of sin(theta)/lambda, where          
-#  lambda is the photon wavelengh and theta is incident angle.               
-#  This function can be approximated by a function:                          
-#                                                                          
-#    f0[k] = c + [SUM a_i*EXP(-b_i*(k^2)) ]                                  
-#                i=1,5                                                       
-#                                                                          
-#  where k = sin(theta) / lambda and c, a_i and b_i                          
-#  are the coefficients tabulated in this file
-
-  #           a1            a2          a3            a4          a5            c           b1           b2           b3           b4           b5      Coh_xs (n)
-  H    :  [ 0.413048,    0.294953,    0.187491,    0.080701,    0.023736,    0.000049,   15.569946,   32.398468,    5.711404,   61.889874,    1.334118,    1.7568]
-  He   :  [ 0.732354,    0.753896,    0.283819,    0.190003,    0.039139,    0.000487,   11.553918,    4.595831,    1.546299,   26.463964,    0.377523,    1.3400]
-  Li   :  [ 0.974637,    0.158472,    0.811855,    0.262416,    0.790108,    0.002542,    4.334946,    0.342451,   97.102966,  201.363831,    1.409234,    0.4540]
-  Be   :  [ 1.533712,    0.638283,    0.601052,    0.106139,    1.118414,    0.002511,   42.662079,    0.595420,   99.106499,    0.151340,    1.843093,    7.6300]
-  B    :  [ 2.085185,    1.064580,    1.062788,    0.140515,    0.641784,    0.003823,   23.494068,    1.137894,   61.238976,    0.114886,    0.399036,    3.5400]
-  C    :  [ 2.657506,    1.078079,    1.490909,   -4.241070,    0.713791,    4.297983,   14.780758,    0.776775,   42.086842,   -0.000294,    0.239535,    5.5510]
-  N    :  [11.893780,    3.277479,    1.858092,    0.858927,    0.912985,   -11.80490,    0.000158,   10.232723,   30.344690,    0.656065,    0.217287,   11.0100]
-  O    :  [ 2.960427,    2.508818,    0.637853,    0.722838,    1.142756,    0.027014,   14.182259,    5.936858,    0.112726,   34.958481,    0.390240,    4.2320]
-  F    :  [ 3.511943,    2.772244,    0.678385,    0.915159,    1.089261,    0.032557,   10.687859,    4.380466,    0.093982,   27.255203,    0.313066,    4.0170]
[…]
