### 2023-12-15 5242616
Moved generate_nanoparticles to utilties with revisions on generation process, based on #21
