+++ debyecalculator/utility/generate.py
@@ -1,0 +1,311 @@
+import torch
+from torch import cdist
+import numpy as np
+#import pandas as pd
+from ase.io import read
+from ase.build import make_supercell
+from ase.build.tools import sort as ase_sort
+from typing import Union, List
+from collections import namedtuple
+import yaml
+import pkg_resources
+import warnings
+from tqdm import tqdm
+
+NanoParticle = namedtuple('NanoParticle', 'elements size occupancy xyz')
+NanoParticleASE = namedtuple('NanoParticleASe', 'ase_structure np_size')
+NanoParticleASEGraph = namedtuple('NanoParticleASEGraph', 'ase_structure np_size edges distances')
+NanoParticleType = Union[
+    List[NanoParticle],
+    NanoParticle,
+    List[NanoParticleASE],
+    NanoParticleASE,
+    List[NanoParticleASEGraph],
+    NanoParticleASEGraph
+]
+
+def get_default_atoms(
+    atom_type: str,
+    output_type: str = 'number'
+) -> List[str]:
+    """
+    Get the default atoms based on the atom type and output type.
+
+    Parameters:
+    - atom_type (str): The type of atoms to retrieve. Accepts either "metal" or "ligand".
+    - output_type (str): The type of output to retrieve. Accepts either "number" or "symbol". Defaults to "number".
+
+    Returns:
+    - atoms (list): The list of default atoms based on the atom type and output type.
+
+    Raises:
+    - ValueError: If the atom_type is not "metal" or "ligand".
+    - ValueError: If the output_type is not "number" or "symbol".
+    """
+    if atom_type == 'metal':
+        if output_type == 'number':
+            atoms = [
+                3, 4, 5, 11, 12, 13, 14, 19, 20, 21, 22, 23, 24, 
+                25, 26, 27, 28, 29, 30, 31, 32, 33, 37, 38, 39, 
+                40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 
+                52, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 
+                66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 
+                78, 79, 80, 81, 82, 83, 88
+            ]
+        elif output_type == 'symbol':
+            atoms = [
+                'Li', 'Be', 'B', 'Na', 'Mg', 'Al', 'Si', 'K', 
+                'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Fe', 'Co', 
+                'Ni', 'Cu', 'Zn', 'Ga', 'Ge', 'As', 'Rb', 'Sr', 
+                'Y', 'Zr', 'Nb', 'Mo', 'Tc', 'Ru', 'Rh', 'Pd', 
+                'Ag', 'Cd', 'In', 'Sn', 'Sb', 'Te', 'Cs', 'Ba', 
+                'La', 'Ce', 'Pr', 'Nd', 'Pm', 'Sm', 'Eu', 'Gd', 
+                'Tb', 'Dy', 'Ho', 'Er', 'Tm', 'Yb', 'Lu', 'Hf', 
+                'Ta', 'W', 'Re', 'Os', 'Ir', 'Pt', 'Au', 'Hg', 
+                'Tl', 'Pb', 'Bi', 'Ra'
+            ]
+        else:
+            raise ValueError('FAILED: Invalid output_type, accepts only "number" or "symbol"')
+    elif atom_type == 'ligand':
+        if output_type == 'number':
+            atoms = [
+                1, 6, 7, 8, 9, 15, 16, 17, 34, 35, 53
+            ]
+        elif output_type == 'symbol':
+            atoms = [
[…]
