+++ debyecalculator/debye_calculator.py
@@ -22,14 +22,12 @@ from ase.build import make_supercell
 from ase.build.tools import sort as ase_sort
 
 from debyecalculator.utility.profiling import Profiler
+from debyecalculator.utility.generate import generate_nanoparticles
 
 import ipywidgets as widgets
 from IPython.display import display, HTML, clear_output
 from ipywidgets import HBox, VBox, Layout
 from tqdm.auto import tqdm
-
-import collections
-import timeit
 
 StructureTuple = namedtuple('StructureTuple', 'elements size occupancy xyz triu_indices unique_inverse unique_form_factors form_avg_sq structure_inverse')
 IqTuple = namedtuple('IqTuple', 'q i')
@@ -122,7 +120,7 @@ class DebyeCalculator:
         self.r = torch.arange(self.rmin, self.rmax, self.rstep).unsqueeze(-1).to(device=self.device)
 
         # Form factor coefficients
-        with open(pkg_resources.resource_filename(__name__, 'form_factor_coef.yaml'), 'r') as yaml_file:
+        with open(pkg_resources.resource_filename(__name__, 'utility/elements_info.yaml'), 'r') as yaml_file:
             self.FORM_FACTOR_COEF = yaml.safe_load(yaml_file)
 
         # Formfactor retrieval lambda
@@ -239,7 +237,7 @@ class DebyeCalculator:
 
             return check_len and check_type and check_shape
 
-        def parse_elements(elements):
+        def parse_elements(elements, size):
             # Get unique elements and construc form factor stacks
             unique_elements, inverse, counts = np.unique(elements, return_counts=True, return_inverse=True)
 
@@ -262,7 +260,7 @@ class DebyeCalculator:
                 size = xyz.shape[0]
                 xyz = xyz.to(device=self.device, dtype=torch.float32)
                 occupancy = torch.ones(xyz.shape[0]).to(device=self.device, dtype=torch.float32)
-                triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse = parse_elements(elements)
+                triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse = parse_elements(elements, size)
 
                 return StructureTuple(elements, size, occupancy, xyz, triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse)
 
@@ -291,25 +289,28 @@ class DebyeCalculator:
                 except:
                     raise IOError(f'Encountered invalid file format when trying to load structure from {structure_source}')
                     
-                triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse = parse_elements(elements)
+                triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse = parse_elements(elements, size)
 
                 return StructureTuple(elements, size, occupancy, xyz, triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse)
 
             elif ext == 'cif':
                 if radii is not None:
-                    structures = self.generate_nanoparticles(structure_source, radii, disable_pbar=disable_pbar, _lightweight_mode=self._lightweight_mode)[0]
+                    structures = generate_nanoparticles(structure_source, radii, disable_pbar=disable_pbar, _lightweight_mode=self._lightweight_mode)
                     structure_tuple_list = []
-
                     for structure in structures:
-                        elements = structure.get_chemical_symbols()
-                        size = len(elements)
-                        occupancy = torch.ones((size), dtype=torch.float32).to(device=self.device)
-                        xyz = torch.tensor(np.array(structure.get_positions())).to(device=self.device, dtype=torch.float32)
-                
[…]
