### 2023-12-15 faa56d3
Updated doc strings, #21
