+++ debyecalculator/debye_calculator.py
@@ -51,25 +51,8 @@ StructureSourceType = Union[
 
 class DebyeCalculator:
     """
-    Calculate the scattering intensity I(q) through the Debye scattering equation, the Total Scattering Structure Function S(q), 
-    the Reduced Total Scattering Function F(q), and the Reduced Atomic Pair Distribution Function G(r) for a given atomic structure.
-
-
-    Parameters:
-        qmin (float): Minimum q-value for the scattering calculation. Default is 1.0.
-        qmax (float): Maximum q-value for the scattering calculation. Default is 30.0.
-        qstep (float): Step size for the q-values in the scattering calculation. Default is 0.1.
-        qdamp (float): Damping parameter caused by the truncated Q-range of the Fourier transformation. Default is 0.04.
-        rmin (float): Minimum r-value for the pair distribution function (PDF) calculation. Default is 0.0.
-        rmax (float): Maximum r-value for the PDF calculation. Default is 20.0.
-        rstep (float): Step size for the r-values in the PDF calculation. Default is 0.01.
-        rthres (float): Threshold value for exclusion of distances below this value in the scattering calculation. Default is 0.0.
-        biso (float): Debye-Waller isotropic atomic displacement parameter. Default is 0.3.
-        device (str): Device to use for computation (e.g., 'cuda' for GPU or 'cpu' for CPU). Default is 'cuda' if the computer has a GPU.
-        batch_size (int or None): Batch size for computation. If None, the batch size will be automatically set. Default is None.
-        lorch_mod (bool): Flag to enable Lorch modification. Default is False.
-        radiation_type (str): Type of radiation for form factor calculations ('xray' or 'neutron'). Default is 'xray'.
-        profile (bool): Activate profiler. Default is False.
+    Calculate the scattering intensity I(Q) through the Debye scattering equation, the Total Scattering Structure Function S(Q), 
+    the Reduced Total Scattering Function F(Q), and the Reduced Atomic Pair Distribution Function G(r) for a given atomic structure.
     """
 
     def __init__(
@@ -91,6 +74,25 @@ class DebyeCalculator:
         _max_batch_size: int = 4000,
         _lightweight_mode: bool = False,
     ) -> None:
+        """
+        Initialize a DebyeCalculator instance with specified parameters.
+
+        Args:
+            qmin (float): Minimum q-value for the scattering calculation. Default is 1.0.
+            qmax (float): Maximum q-value for the scattering calculation. Default is 30.0.
+            qstep (float): Step size for the q-values in the scattering calculation. Default is 0.1.
+            qdamp (float): Damping parameter caused by the truncated Q-range of the Fourier transformation. Default is 0.04.
+            rmin (float): Minimum r-value for the pair distribution function (PDF) calculation. Default is 0.0.
+            rmax (float): Maximum r-value for the PDF calculation. Default is 20.0.
+            rstep (float): Step size for the r-values in the PDF calculation. Default is 0.01.
+            rthres (float): Threshold value for exclusion of distances below this value in the scattering calculation. Default is 0.0.
+            biso (float): Debye-Waller isotropic atomic displacement parameter. Default is 0.3.
+            device (str): Device to use for computation (e.g., 'cuda' for GPU or 'cpu' for CPU). Default is 'cuda' if the computer has a GPU.
+            batch_size (int or None): Batch size for computation. If None, the batch size will be automatically set. Default is None.
+            lorch_mod (bool): Flag to enable Lorch modification. Default is False.
+            radiation_type (str): Type of radiation for form factor calculations ('xray' or 'neutron'). Default is 'xray'.
+            profile (bool): Activate profiler. Default is False.
+        """
             
         # Set parameters
         self.qmin = qmin
@@ -155,7 +157,10 @@ class DebyeCalculator:
         self,
     ) -> None:
[…]
