+++ debyecalculator/utility/generate.py
@@ -1,7 +1,6 @@
 import torch
 from torch import cdist
 import numpy as np
-#import pandas as pd
 from ase.io import read
 from ase.build import make_supercell
 from ase.build.tools import sort as ase_sort
@@ -42,43 +41,41 @@ def get_default_atoms(
     - ValueError: If the atom_type is not "metal" or "ligand".
     - ValueError: If the output_type is not "number" or "symbol".
     """
+    METAL_SYMBOLS = [
+        'Li', 'Be', 'B', 'Na', 'Mg', 'Al', 'Si', 'K',
+        'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Fe', 'Co',
+        'Ni', 'Cu', 'Zn', 'Ga', 'Ge', 'As', 'Rb', 'Sr',
+        'Y', 'Zr', 'Nb', 'Mo', 'Tc', 'Ru', 'Rh', 'Pd',
+        'Ag', 'Cd', 'In', 'Sn', 'Sb', 'Te', 'Cs', 'Ba',
+        'La', 'Ce', 'Pr', 'Nd', 'Pm', 'Sm', 'Eu', 'Gd',
+        'Tb', 'Dy', 'Ho', 'Er', 'Tm', 'Yb', 'Lu', 'Hf',
+        'Ta', 'W', 'Re', 'Os', 'Ir', 'Pt', 'Au', 'Hg',
+        'Tl', 'Pb', 'Bi', 'Ra'
+    ]
+
+    LIGAND_SYMBOLS = ['H', 'C', 'N', 'O', 'F', 'P', 'S', 'Cl', 'Se', 'Br', 'I']
+
+    METAL_NUMBERS = [
+        3, 4, 5, 11, 12, 13, 14, 19, 20, 21, 22, 23, 24,
+        25, 26, 27, 28, 29, 30, 31, 32, 33, 37, 38, 39,
+        40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51,
+        52, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65,
+        66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77,
+        78, 79, 80, 81, 82, 83, 88
+    ]
+
+    LIGAND_NUMBERS = [1, 6, 7, 8, 9, 15, 16, 17, 34, 35, 53]
+
     if atom_type == 'metal':
-        if output_type == 'number':
-            atoms = [
-                3, 4, 5, 11, 12, 13, 14, 19, 20, 21, 22, 23, 24, 
-                25, 26, 27, 28, 29, 30, 31, 32, 33, 37, 38, 39, 
-                40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 
-                52, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 
-                66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 
-                78, 79, 80, 81, 82, 83, 88
-            ]
-        elif output_type == 'symbol':
-            atoms = [
-                'Li', 'Be', 'B', 'Na', 'Mg', 'Al', 'Si', 'K', 
-                'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Fe', 'Co', 
-                'Ni', 'Cu', 'Zn', 'Ga', 'Ge', 'As', 'Rb', 'Sr', 
-                'Y', 'Zr', 'Nb', 'Mo', 'Tc', 'Ru', 'Rh', 'Pd', 
-                'Ag', 'Cd', 'In', 'Sn', 'Sb', 'Te', 'Cs', 'Ba', 
[…]
