### 2023-12-15 c2415ce
Added support for (atomic_numbers (Int), xyz_coordinates) as input
