### 2023-12-15 FrederikLizakJohansen/DebyeCalculator#21 @FrederikLizakJohansen
Updates for the current issue (check out the [issue-21 branch](https://github.com/FrederikLizakJohansen/DebyeCalculator/tree/issue-21)):

- Improved `iq`, `sq`, `fq`, `gr`, and `_get_all `functions. They can now handle input in the form of tuples with atoms listed as either numbers or element names.

- Changed the output of `iq`, `sq`, `fq`, `gr`, and `_get_all `to use namedtuples with more meaningful names. This tweak was based on feedback from issue #20.

- Moved the `generate_nanoparticles `function to the utility folder as a standalone function and updated it.

I'll be adding more unit tests before merging everything with the main branch.
