### 2023-12-18 2df4aba
Higher precision on Interact outputs, minor fixes in tuple-input
