### 2023-12-18 ab3a3e2
Added unit tests for tuple input as well as new standalone generation function, #21
