+++ debyecalculator/test_debye_calculator.py
@@ -1,6 +1,19 @@
 import pytest, torch
 from debyecalculator import DebyeCalculator
+from debyecalculator.utility.generate import generate_nanoparticles
 import numpy as np
+from ase.io import read
+import pkg_resources
+import yaml
+
+# Elements to atomic numbers map
+with open(pkg_resources.resource_filename(__name__, 'utility/elements_info.yaml'), 'r') as yaml_file:
+    element_info = yaml.safe_load(yaml_file)
+element_to_atomic_number = {}
+for i, (key, value) in enumerate(element_info.items()):
+    if i > 97:
+        break
+    element_to_atomic_number[key] = value[12]
 
 def test_init_defaults():
     calc = DebyeCalculator()
@@ -19,46 +32,152 @@ def test_init_defaults():
     assert calc.radiation_type == 'xray', f"Expected radiation_type to be 'xray', but got {calc.radiation_type}"
     assert calc.profile == False, f"Expected profile to be False, but got {calc.profile}"
 
-def test_iq():
+def test_iq_xyz():
     # Load the expected scattering intensity from a file
     ph = np.loadtxt('debyecalculator/unittests_files/icsd_001504_cc_r6_lc_2.85_6_tetragonal_Iq.dat')
-    Q_expected, Iq_expected = ph[:,0], ph[:,1]
+    q_expected, iq_expected = ph[:,0], ph[:,1]
 
     # Calculate the scattering intensity using the DebyeCalculator
     calc = DebyeCalculator()
-    Q, Iq = calc.iq('debyecalculator/unittests_files/icsd_001504_cc_r6_lc_2.85_6_tetragonal.xyz')
+    q, iq = calc.iq('debyecalculator/unittests_files/icsd_001504_cc_r6_lc_2.85_6_tetragonal.xyz')
 
     # Check that the calculated scattering intensity matches the expected value
-    assert np.allclose(Q, Q_expected, atol=1e-05, rtol=1e-03), f"Expected Q to be {Q_expected}, but got {Q}"
-    assert np.allclose(Iq, Iq_expected, atol=1e-05, rtol=1e-03), f"Expected Iq to be {Iq_expected}, but got {Iq}"
+    assert np.allclose(q, q_expected, atol=1e-04, rtol=1e-03), f"Expected Q to be {q_expected}, but got {q}"
+    assert np.allclose(iq, iq_expected, atol=1e-04, rtol=1e-03), f"Expected I(Q) to be {iq_expected}, but got {iq}"
 
-def test_sq():
+def test_iq_cif():
+
+    # Load expected scattering intensity generated from cif
+    ph = np.genfromtxt('debyecalculator/unittests_files/iq_AntiFluorite_Co2O_radius10.0.dat', delimiter=',', skip_header=15)
+    q_expected, iq_expected = ph[:,0], ph[:,1]
+
+    # Calculate the scatering intensity using DebyeCalculator
+    calc = DebyeCalculator(qstep=0.05)
+    q, iq = calc.iq('data/AntiFluorite_Co2O.cif', radii=10.0)
+    
+    # Check that the calculated scattering intensity matches the expected value
+    assert np.allclose(q, q_expected, atol=1e-04, rtol=1e-03), f"Expected Q to be {q_expected}, but got {q}"
+    assert np.allclose(iq, iq_expected, atol=1e-04, rtol=1e-03), f"Expected I(Q) to be {iq_expected}, but got {iq}"
+
+def test_iq_tuple():
+
+    # Load expected scattering intensity
[…]
