### 2023-12-18 FrederikLizakJohansen/DebyeCalculator#21 @FrederikLizakJohansen
The changes have now been merged to the main branch and the functions are working as [intented](https://github.com/FrederikLizakJohansen/DebyeCalculator/actions/runs/7245417561).

I am closing the issue here. Thank you again for the suggestion.
