### 2023-12-18 FrederikLizakJohansen/DebyeCalculator#20 @FrederikLizakJohansen
Thank you for the suggestion!
This issue has been adressed and resolved in #21. Therefore I am closing this issue. Feel free to reopen or create a new issue, if the current implementation can be further improved.
