### 2023-12-19 25ffe21
Explicit handling of CUDA availability to the users with UserWarning, based on #26
