### 2023-12-19 a52c365
Updated "Example Usage" according to recent changes in function-input
