### 2023-12-19 032652d
Added section on prerequisites and GPU acceleration support
