### 2023-12-19 bf1280a
Removed Pytorch installation, see README instead
