+++ pyproject.toml
@@ -28,10 +28,6 @@ matplotlib = [
   {version = "^3.4.3", python = "^3.7,<3.9"},
   {version = "^3.8.0", python = "^3.9,<3.12"}
 ]
-torch = [
-  {version = "1.13.1", python = "^3.7,<3.9"},
-  {version = "^2.0.1", python = "^3.9,<3.12"}
-]
 pyyaml = "^6.0.1"
 tqdm = "^4.66.1"
 ase = "^3.22.1"
