### 2023-12-19 2131bb5
Added try/except for detecting importerror on pytorch
