### 2023-12-19 4a0a173
Added global calculator so as to not overflood the user with warnings
