+++ debyecalculator/test_debye_calculator.py
@@ -15,8 +15,11 @@ for i, (key, value) in enumerate(element_info.items()):
         break
     element_to_atomic_number[key] = value[12]
 
+# Global calculator object
+calc = DebyeCalculator()
+
 def test_init_defaults():
-    calc = DebyeCalculator()
+    #calc = DebyeCalculator()
     assert calc.qmin == 1.0, f"Expected qmin to be 1.0, but got {calc.qmin}"
     assert calc.qmax == 30.0, f"Expected qmax to be 30.0, but got {calc.qmax}"
     assert calc.qstep == 0.1, f"Expected qstep to be 0.1, but got {calc.qstep}"
@@ -38,7 +41,8 @@ def test_iq_xyz():
     q_expected, iq_expected = ph[:,0], ph[:,1]
 
     # Calculate the scattering intensity using the DebyeCalculator
-    calc = DebyeCalculator()
+    #calc = DebyeCalculator()
+    calc.update_parameters(qstep=0.1)
     q, iq = calc.iq('debyecalculator/unittests_files/icsd_001504_cc_r6_lc_2.85_6_tetragonal.xyz')
 
     # Check that the calculated scattering intensity matches the expected value
@@ -52,7 +56,8 @@ def test_iq_cif():
     q_expected, iq_expected = ph[:,0], ph[:,1]
 
     # Calculate the scatering intensity using DebyeCalculator
-    calc = DebyeCalculator(qstep=0.05)
+    calc.update_parameters(qstep=0.05)
+    #calc = DebyeCalculator(qstep=0.05)
     q, iq = calc.iq('data/AntiFluorite_Co2O.cif', radii=10.0)
     
     # Check that the calculated scattering intensity matches the expected value
@@ -73,7 +78,8 @@ def test_iq_tuple():
     structure_tuple_str = (ase_elements, ase_xyz)
     structure_tuple_int = (np.array([element_to_atomic_number[e] for e in ase_elements]), ase_xyz)
 
-    calc = DebyeCalculator(qstep=0.05)
+    #calc = DebyeCalculator(qstep=0.05)
+    calc.update_parameters(qstep=0.05)
     q_str, iq_str = calc.iq(structure_tuple_str)
     q_int, iq_int = calc.iq(structure_tuple_int)
 
@@ -89,7 +95,8 @@ def test_sq_xyz():
     q_expected, sq_expected = ph[:,0], ph[:,1]
 
     # Calculate the structure factor using the DebyeCalculator
-    calc = DebyeCalculator()
+    #calc = DebyeCalculator()
+    calc.update_parameters(qstep=0.1)
     q, sq = calc.sq('debyecalculator/unittests_files/icsd_001504_cc_r6_lc_2.85_6_tetragonal.xyz')
 
     # Check that the calculated structure factor matches the expected value
@@ -102,7 +109,8 @@ def test_sq_cif():
     q_expected, sq_expected = ph[:,0], ph[:,1]
 
     # Calculate the structure factor using the DebyeCalculator
-    calc = DebyeCalculator(qstep=0.05)
+    #calc = DebyeCalculator(qstep=0.05)
+    calc.update_parameters(qstep=0.05)
     q, sq = calc.sq('data/AntiFluorite_Co2O.cif', radii=10.0)
 
     # Check that the calculated structure factor matches the expected value
@@ -121,7 +129,8 @@ def test_sq_tuple():
     structure_tuple_str = (ase_elements, ase_xyz)
     structure_tuple_int = (np.array([element_to_atomic_number[e] for e in ase_elements]), ase_xyz)
 
-    calc = DebyeCalculator(qstep=0.05)
+    #calc = DebyeCalculator(qstep=0.05)
[…]
