### 2023-12-19 d90ba98
Added tests for CPU-only versions
