### 2023-12-19 8c4f38a
Added tests for CPU-only versions
