### 2023-12-19 FrederikLizakJohansen/DebyeCalculator#27 @AndySAnker
Thank you for the suggestion of changing the naming of the variables.

We have now renamed `structure_path` to `structure_source`. This change is more in line with the expanded input formats that were developed after the feedback in suggestion #21.

> All the computation return tuples of parameters in reverse order from the name of the method. Is this on purpose?

Yes, that's intentional. The functions return Q and I (in that order) since it's customary to plot Q on the x-axis. However, we usually refer to it as Iq in discussions. Therefore we named the function `calc.iq`. This naming follows the convention used in DiffPy-CMI. To address your `namedtuples` point, we have updated the output so it can be accessed in two ways:

Directly:
`q, iq = calc.iq(...)`

Via namedtuple:
`scattering = calc.iq(...)`
`q = scattering.q`
`iq = scattering.i`

I am closing this issue. Feel free to reopen the issue if you would like to see any further changes made to this.
