### 2023-12-19 bb99c3a
Added tests for CPU-only versions
