+++ .github/workflows/unit-tests.yml
@@ -42,6 +42,7 @@ jobs:
         else
           echo "Unsupported value for device"
           exit 1
+        fi
     - name: Install DebyeCalculator dependencies
       run: |
         poetry install
