### 2023-12-19 c95df49
Added tests for CPU/CUDA
