### 2023-12-19 93224b9
Added tests for CPU/CUDA
