+++ .github/workflows/unit-tests.yml
@@ -9,11 +9,11 @@ on:
       - main
 
 jobs:
-  build:
+  build-3:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v2
@@ -30,21 +30,7 @@ jobs:
         poetry env use python
     - name: Install PyTorch CPU
       run: |
-        if [ "${{ matrix.python-version }}" == "3.7"]; then
-          poetry run pip install torch==1.13.1+cpu torchvision==0.14.1+cpu torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cpu
-        fi
-        if [ "${{ matrix.python-version }}" == "3.8"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.9"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.10"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.11"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
+        poetry run pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
     - name: Install DebyeCalculator dependencies
       run: |
         poetry install
@@ -57,21 +43,49 @@ jobs:
         poetry env use python
     - name: Install PyTorch CUDA
       run: |
-        if [ "${{ matrix.python-version }}" == "3.7"]; then
-          poetry run pip install torch==1.13.1+cu116 torchvision==0.14.1+cu116 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu116
-        fi
-        if [ "${{ matrix.python-version }}" == "3.8"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.9"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.10"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
-        if [ "${{ matrix.python-version }}" == "3.11"]; then
-          poetry run pip install torch torchvision torchaudio
-        fi
+        poetry run pip install torch torchvision torchaudio
+    - name: Install DebyeCalculator dependencies
+      run: |
+        poetry install
+    - name: Run pytest CUDA
+      run: |
+        poetry run pytest
+
+  build-37:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.7"]
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Setup poetry
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry
+    - name: Reset poetry env
+      run: |
+        poetry env use python
+    - name: Install PyTorch CPU
+      run: |
+        poetry run pip install torch==1.13.1+cpu torchvision==0.14.1+cpu torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cpu
+    - name: Install DebyeCalculator dependencies
+      run: |
+        poetry install
+    - name: Run pytest CPU
+      run: |
+        poetry run pytest
+    - name: Reset poetry env
+      run: |
[…]
