### 2023-12-19 FrederikLizakJohansen/DebyeCalculator#24 @FrederikLizakJohansen
Thank you for bringing this to our attention. This is something that we've naturally been concerned about as well. We've made updates to the installation instructions and provided more information about the prerequisites for using GPU support in the DebyeCalculator package, as well as implemented more explicit warnings about CUDA availability into the software. With these changes, PyTorch will not be installed with the DebyeCalculator package within the pip-installation, but instead be a prerequisite to the package. To sum up:

1. **PyTorch Dependency:** We want to clarify that DebyeCalculator depends on PyTorch, and users need to install PyTorch separately before using our package. We've updated the README to include PyTorch as a prerequisite. Users are now directed to install PyTorch from the official website before proceeding with DebyeCalculator installation.

2. **More Specific Error Handling:** In response to your feedback, we've enhanced the error/warning messages. If PyTorch is not installed, users will now receive a more informative message, guiding them to install PyTorch with a direct link to the official PyTorch installation guide as well as this Github page. The README now also contains a dedicated "GPU Support" section that provides steps for verifying CUDA availability.

These changes will be applied to **DebyeCalculator v.1.0.4 and onwards through the PyPI installation (https://pypi.org/project/debyecalculator/)**

We hope these enhancements address your concerns. If you encounter any further issues or have additional suggestions, don't hesitate to let us know, so we can reopen the issue.
