### 2023-12-19 FrederikLizakJohansen/DebyeCalculator#26 @FrederikLizakJohansen
Thank you for bringing this to our attention. This issue has been resolved in #24.

Closing this issue.
