### 2023-12-21 c7bd105
Fixes for device specification
