### 2023-12-21 94fddd0
Overhaul of upload/download mechanism for Interact() function
