+++ debyecalculator/debye_calculator.py
@@ -1,6 +1,11 @@
 import os
+import tempfile
+from io import BytesIO
+import zipfile
 import sys
 import base64
+import hashlib
+import time
 import yaml
 import pkg_resources
 import warnings
@@ -37,6 +42,7 @@ from debyecalculator.utility.generate import generate_nanoparticles
 import ipywidgets as widgets
 from IPython.display import display, HTML, clear_output
 from ipywidgets import HBox, VBox, Layout
+from functools import partial
 from tqdm.auto import tqdm
 
 StructureTuple = namedtuple('StructureTuple', 'elements size occupancy xyz triu_indices unique_inverse unique_form_factors form_avg_sq structure_inverse')
@@ -69,7 +75,7 @@ class DebyeCalculator:
         self,
         qmin: float = 1.0,
         qmax: float = 30.0,
-        qstep: float = 0.1,
+        qstep: float = 0.05,
         qdamp: float = 0.04,
         rmin: float = 0.0,
         rmax: float = 20.0,
@@ -351,7 +357,7 @@ class DebyeCalculator:
 
             elif ext == 'cif':
                 if radii is not None:
-                    structures = generate_nanoparticles(structure_source, radii, disable_pbar=disable_pbar, _lightweight_mode=self._lightweight_mode)
+                    structures = generate_nanoparticles(structure_source, radii, disable_pbar=disable_pbar, _lightweight_mode=self._lightweight_mode, device=self.device)
                     structure_tuple_list = []
                     for structure in structures:
                         triu_indices, unique_inverse, unique_form_factors, form_avg_sq, structure_inverse = parse_elements(structure.elements, structure.size)
@@ -932,59 +938,32 @@ class DebyeCalculator:
         # File selection sizes
         header_widths = [105*1.8, 130*1.8]
         header_widths = [str(i)+'px' for i in header_widths]
-
-        # Folder selection
-        folder = widgets.Text(description='', placeholder='Enter data directory', disabled=False, layout=Layout(width='650px'))
         
-        # Dropdown file sections
-        DEFAULT_MSGS = ['No valid files in entered directory', 'Select data file']
-        select_file_1 = widgets.Dropdown(options=DEFAULT_MSGS, value=DEFAULT_MSGS[0], disabled=True, layout=Layout(width='650px'))
-        select_file_2 = widgets.Dropdown(options=DEFAULT_MSGS, value=DEFAULT_MSGS[0], disabled=True, layout=Layout(width='650px'))
-        
-        # File 1
-        select_file_desc_1 = HBox([widgets.Image(value=file_1_img, format='png', layout=Layout(object_fit='contain', object_position='20px', width='32px'))], layout=Layout(width='88px'))
-        select_radius_desc_1 = HBox([widgets.Image(value=radius_a_img, format='png', layout=Layout(object_fit='contain', object_position='20px', width='60px', visibility='hidden'))], layout=Layout(width='88px'))
-        select_radius_1 = widgets.FloatText(min = 0, max = 50, step=0.01, value=5, disabled = False, layout = Layout(width='50px', visibility='hidden'))
-        cif_text_1 = widgets.Text(
-            value='Given radius, generate spherical nanoparticles (NP) from crystallographic information files (CIFs)', 
-            disabled=True,
-            layout=Layout(width='595px', visibility='hidden')
-        )
+        # Upload files
+        # Specify the allowed file extensions
+        allowed_extensions = ['.cif', '.xyz']
 
-        # File 2
-        select_file_desc_2 = HBox([widgets.Image(value=file_2_img, format='png', layout=Layout(object_fit='contain', object_position='20px', width='32px'))], layout=Layout(width='88px'))
[…]
