### 2023-12-21 fe41102
Added possibility for loading example data into interact()
