### 2024-01-02 FrederikLizakJohansen/DebyeCalculator#15 @FrederikLizakJohansen
Thank you for bringing this to our attention, and we appreciate your patience. (Happy New Year!)

We have investigated the issue you reported regarding the "generate plot" button not triggering the simulation in Google Colab. To combat this issue, we have updated the "interact" function so that it is now possible to upload and download data-files more easily. This, however, have introduced some additional challenges. Colab is now flagging the "interact" widget as a "third-party-widget," causing the generate plot button to malfunction. Currently, the only known solution is to downgrade the "Ipywidgets" version, but this may lead to other components breaking.

In light of these difficulties, we have decided to reserve the implementation of the "interact" function on HuggingFace for a future version. Meanwhile, we recommend users to pip install versions <1.0.5 of the DebyeCalculator specifically for Colab usage, where data-files (xyz or cif) can be manually uploaded to Colab, and the widget can access them using the now deprecated path string option. If you want to test this out, you can clone the repository and manually upload the "data" folder to the Colab page. You should then be able to generate the corresponding plots shown on the figure in the paper.

We understand that this workaround is not ideal, and we apologize for any inconvenience caused. We are actively working on addressing this issue and appreciate your understanding. We will keep this issue open until we find a sustainable solution. If you have any further questions or concerns, please feel free to reach out.

Thank you again for your contribution and understanding.
