### 2024-01-04 05f976c
minor fixes to nanoparticle distances for graph generation
