### 2024-01-04 fd65a84
minor fixes to nanoparticle distances for graph generation
