### 2024-01-04 a69ac8c
minor fixes to nanoparticle distances for graph generation
