### 2024-01-04 dc28b1f
minor fixes to nanoparticle distances for graph generation
