### 2024-01-06 fe642fa
Minor performance fixes in graph generation
