### 2024-01-12 FrederikLizakJohansen/DebyeCalculator#15 @FrederikLizakJohansen
Thanks again for the suggestion. We've incorporated the link with a brief message in the Colab notebook.
