### 2024-01-17 e9e9181
added benchmarking class
