+++ debyecalculator/utility/benchmark.py
@@ -1,0 +1,246 @@
+import tempfile
+import os
+import csv
+import pkg_resources
+import argparse
+import warnings
+import torch
+from time import time
+import numpy as np
+from tqdm.auto import tqdm, trange
+import matplotlib.pyplot as plt
+from ase.io import write, read
+from debyecalculator import DebyeCalculator
+from debyecalculator.utility.generate import generate_nanoparticles
+from prettytable import PrettyTable, from_csv
+
+from typing import Union, List, Any
+from collections import namedtuple
+
+class Statistics:
+    def __init__(self, means, stds, name, radii, num_atoms, cuda_mem, device, batch_size):
+
+        self.means = means
+        self.stds = stds
+        self.name = name
+        self.radii = radii
+        self.num_atoms = num_atoms
+        self.cuda_mem = cuda_mem
+        self.device = device
+        self.batch_size = batch_size
+        
+        # Create table
+        self.table_fields = ['Radius [Å]', 'Num. atoms', 'Mean [s]', 'Std [s]', 'MaxAlloc. CUDA mem [MB]']
+        self.pt = PrettyTable(self.table_fields)
+        self.pt.align = 'r'
+        self.pt.padding_width = 1
+        self.pt.title = 'Benchmark Generator / ' + self.device + ' / Batch Size: ' + str(self.batch_size)
+        self.data = [[str(float(r)), str(int(n)), f'{m:1.5f}', f'{s:1.5f}', f'{c:1.5f}'] for r,n,m,s,c in zip(self.radii, list(num_atoms), list(means), list(stds), list(cuda_mem))]
+        for d in self.data:
+            self.pt.add_row(d)
+
+    def __str__(self):
+        return str(self.pt)
+
+    def __repr__(self):
+        return f'Statistics (\n\tname = {self.name},\n\tradii = {self.radii},\n\tnum_atoms = {self.num_atoms},\n\tmeans = {self.means},\n\tstds = {self.stds},\n\tcuda_mem = {self.cuda_mem},\n\tbatch_size = {self.batch_size},\n\tdevice = {self.device}\n)'
+
+class DebyeBenchmarker:
+    def __init__(
+        self,
+        radii: Union[List, np.ndarray, torch.Tensor] = [5],
+        show_progress_bar: bool = True,
+        custom_cif: str = None,
+        **kwargs,
+    ) -> None:
+
+        self.set_radii(list(radii))
+        self.cif = pkg_resources.resource_filename(__name__, 'benchmark_structure.cif')
+        self.custom_cif = custom_cif
+
+        self.show_progress_bar = show_progress_bar
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            self.debye_calc = DebyeCalculator(**kwargs)
+
+    def set_debye_parameters(self, **debye_parameters: Any):
+        self.debye_calc.update_parameters(debye_parameters)
+
+    def set_device(self, device: str):
+        self.debye_calc.update_parameters(device=device)
+
+    def set_batch_size(self, batch_size: int):
+        self.debye_calc.update_parameters(batch_size=batch_size)
+
+    def set_radii(self, radii: Union[List, np.ndarray, torch.Tensor]):
+        self.radii = list(radii)
+    
+    def benchmark_generation(
+        self,
+        individually: bool = False,
+        repetitions: int = 1,
+        csv_output: Union[None, str] = None,
+    ) -> Statistics:
+        
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+
+            # Create metrics arrays
+            means = np.zeros(len(self.radii))
+            stds = np.zeros(len(self.radii))
+            num_atoms = np.zeros(len(self.radii))
+            cuda_mem = np.zeros(len(self.radii))
+            
[…]
