+++ runtime_tests/produce_figure_data.py
@@ -1,17 +1,24 @@
 import tempfile
 import os
+import pkg_resources
 import argparse
 import torch
 from time import time
 import numpy as np
 from tqdm.auto import tqdm, trange
 import matplotlib.pyplot as plt
-from ase.io import write
-from diffpy.structure import loadStructure
-from diffpy.srreal.pdfcalculator import DebyePDFCalculator
+from ase.io import write, read
 from debyecalculator import DebyeCalculator
+from debyecalculator.utility.generate import generate_nanoparticles
+from prettytable import PrettyTable, from_csv
+from debyecalculator import StructureSourceType
 
 def time_methods(args):
+
+    if args['gen_diffpy']:
+        from diffpy.structure import loadStructure
+        from diffpy.srreal.pdfcalculator import DebyePDFCalculator
+    
     print('Generating data for timing')
     structure_path = args['cif']
     radii = list(np.arange(args['min_radius'], args['max_radius'], args['step_radius']))
@@ -25,32 +32,57 @@ def time_methods(args):
     def time_debye_calculator(device, batch_size, output_folder):
         mu, sigma = [], []
         debye_calc = DebyeCalculator(device=device, qmin=1, qmax=25, qstep=0.1, biso=0.3, batch_size=batch_size)
-        nps, sizes = debye_calc.generate_nanoparticles(structure_path, radii)
-        n_atoms = [len(np) for np in nps]
+        nps = generate_nanoparticles(structure_path, radii, _reverse_order=False)
+        sizes = [r * 2 for r in radii]
+        n_atoms = [nano.size for nano in nps]
 
-        for i in trange(len(radii), leave=False):
-            with tempfile.TemporaryDirectory() as tmpdirname:
-                tmp_structure_path = os.path.join(tmpdirname,'tmp_struc.xyz')
-                write(tmp_structure_path, nps[i], 'xyz')
-
-                timings = []
-                for j in range(args['reps']+2):
-                    t = time()
-                    debye_calc.gr(tmp_structure_path);
-                    t = time() - t
-                    if j > 1: 
-                        timings.append(t)
+        pbar = tqdm(desc='Benchmarking...', total=len(radii))
+        for i,nano in enumerate(nps):
+            timings = []
+            for j in range(args['reps']+2):
+                t = time()
+                debye_calc.gr((nano.elements, nano.xyz))
+                t = time() - t
+                if j > 1:
+                    timings.append(t)
             mu.append(np.mean(timings))
             sigma.append(np.std(timings))
-        
-            if device == 'cpu':
-                out = np.array([sizes[:(i+1)], n_atoms[:(i+1)], np.array(mu), np.array(sigma)]).T
-                np.savetxt(os.path.join(output_folder,f'timings_cpu_bs{batch_size}.csv'), out, delimiter=',', header='diameter, n_atoms, mu, sigma', fmt='%f')
-            elif device == 'cuda':
-                out = np.array([sizes[:(i+1)], n_atoms[:(i+1)], np.array(mu), np.array(sigma)]).T
-                gpu_id = torch.cuda.get_device_name()
-                np.savetxt(os.path.join(output_folder,f'timings_cuda_{gpu_id}_bs{batch_size}.csv'), out, delimiter=',', header='diameter, n_atoms, mu, sigma', fmt='%f')
+            pbar.update(1)
+        pbar.close()
 
+        return sizes, n_atoms, mu, sigma
+
+    def save_csv(sizes, n_atoms, means, stds, batch_size, device):
+        if device.lower() == 'cpu':
+            name = f'timings_cpu_bs{batch_size}.csv'
[…]
