### 2024-01-17 80ef26b
Update README.md
Added partial Gr documentation
