### 2024-01-17 FrederikLizakJohansen/DebyeCalculator#22 @AndySAnker
Thank you for the great suggestion.

With #20 we have added the possibility of providing the structure as a tuple. This means that calculating the partial Gr has become very easy:

```python
# Create a mask for oxygen atoms
mask = [atom == "O" for atom in structure_tuple[0]]

# Apply the mask to the structure_tuple
structure_tuple = (
    [atom for atom, m in zip(structure_tuple[0], mask) if m],
    structure_tuple[1][mask]
)

# Calculate Pair Distribution Function
r_partial, Gr_partial = calc.gr(structure_source=structure_tuple)
```

We have also added this example to the readme file.

I am closing this issue here. Feel free to reopen the issue if you would like to see any further changes made to this.
