### 2024-01-17 5897264
added doc strings and missing type annotations
