+++ debyecalculator/utility/benchmark.py
@@ -18,7 +18,43 @@ from typing import Union, List, Any
 from collections import namedtuple
 
 class Statistics:
-    def __init__(self, means, stds, name, radii, num_atoms, cuda_mem, device, batch_size):
+    """
+    A class to store and represent benchmark statistics.
+
+    Attributes:
+    - means (List[float]): List of mean values.
+    - stds (List[float]): List of standard deviation values.
+    - name (str): Name of the statistics.
+    - radii (List[float]): List of radii.
+    - num_atoms (List[int]): List of number of atoms.
+    - cuda_mem (List[float]): List of CUDA memory values.
+    - device (str): Device used for benchmarking.
+    - batch_size (int): Batch size used for benchmarking.
+    """
+    def __init__(
+        self,
+        means: List[float],
+        stds: List[float],
+        name: str,
+        radii: List[float],
+        num_atoms: List[int],
+        cuda_mem: List[float],
+        device: str,
+        batch_size: int
+    ) -> None:
+        """
+        Initialize Statistics with benchmarking results.
+
+        Parameters:
+        - means (List[float]): List of mean values.
+        - stds (List[float]): List of standard deviation values.
+        - name (str): Name of the statistics.
+        - radii (List[float]): List of radii.
+        - num_atoms (List[int]): List of number of atoms.
+        - cuda_mem (List[float]): List of CUDA memory values.
+        - device (str): Device used for benchmarking.
+        - batch_size (int): Batch size used for benchmarking.
+        """
 
         self.means = means
         self.stds = stds
@@ -39,13 +75,29 @@ class Statistics:
         for d in self.data:
             self.pt.add_row(d)
 
-    def __str__(self):
+    def __str__(self) -> str:
+        """
+        Return a PrettyTable Statistics table.
+        """
         return str(self.pt)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
+        """
+        Return a detailed string representation of the Statistics object.
+        """
         return f'Statistics (\n\tname = {self.name},\n\tradii = {self.radii},\n\tnum_atoms = {self.num_atoms},\n\tmeans = {self.means},\n\tstds = {self.stds},\n\tcuda_mem = {self.cuda_mem},\n\tbatch_size = {self.batch_size},\n\tdevice = {self.device}\n)'
 
 class DebyeBenchmarker:
+    """
+    A class for benchmarking Debye calculations.
+
+    Attributes:
+    - radii (List[float]): List of radii for benchmarking.
+    - cif (str): Path to reference CIF file used for benchmarking.
+    - custom_cif (str): Custom CIF file path (if provided).
+    - show_progress_bar (bool): Flag to control progress bar display.
+    - debye_calc (DebyeCalculator): Debye calculator instance.
+    """
     def __init__(
         self,
         radii: Union[List, np.ndarray, torch.Tensor] = [5],
@@ -53,6 +105,15 @@ class DebyeBenchmarker:
         custom_cif: str = None,
         **kwargs,
     ) -> None:
+        """
+        Initialize DebyeBenchmarker.
+
+        Parameters:
+        - radii (Union[List, np.ndarray, torch.Tensor]): List of radii for benchmarking.
+        - show_progress_bar (bool): Flag to control progress bar display.
+        - custom_cif (str): Custom CIF file path (if provided).
+        - **kwargs: Additional keyword arguments for DebyeCalculator.
+        """
 
         self.set_radii(list(radii))
         self.cif = pkg_resources.resource_filename(__name__, 'benchmark_structure.cif')
@@ -63,24 +124,57 @@ class DebyeBenchmarker:
             warnings.simplefilter("ignore")
[…]
