### 2024-01-17 FrederikLizakJohansen/DebyeCalculator#14 @AndySAnker
Thank you for the great suggestion of using sphinx to generate an API and publish it in readthedoc.

We have done this (https://debyecalculator.readthedocs.io/en/latest/) while maintaining step-by-step use of 'DebyeCalculator' in the readme file of the project.

We have also added badges to the readme file of the pypi, python version, license, ChemRxiv file and readthedocs.

I will close this issue - let us know if it should be reopened.
