### 2024-01-17 FrederikLizakJohansen/DebyeCalculator#13 @AndySAnker
Thank you for making us aware that version tag and release was not included as part of the package.

We have now added a version tag to the readme file and the published package (https://pypi.org/project/debyecalculator/)

We are also currently working on publishing a release on GitHub. More about that in the following days.
