### 2024-01-17 #6024 @AndySAnker
Thanks for the great feedback @KedoKudo @marcocamma @elena-pascal. We are currently finalising the new release of DebyeCalculator and will provide a more detailed response here at the end of this week/beginning of next week.

Thanks for your patience.
