### 2024-01-21 e928d2a
simplified benchmarker class
