+++ debyecalculator/utility/benchmark.py
@@ -38,7 +38,8 @@ class Statistics:
         name: str,
         radii: List[float],
         num_atoms: List[int],
-        cuda_mem: List[float],
+        cuda_mem_structure: List[float],
+        cuda_mem_calculations: List[float],
         device: str,
         batch_size: int
     ) -> None:
@@ -61,17 +62,18 @@ class Statistics:
         self.name = name
         self.radii = radii
         self.num_atoms = num_atoms
-        self.cuda_mem = cuda_mem
+        self.cuda_mem_structure = cuda_mem_structure
+        self.cuda_mem_calculations = cuda_mem_calculations
         self.device = device
         self.batch_size = batch_size
         
         # Create table
-        self.table_fields = ['Radius [Å]', 'Num. atoms', 'Mean [s]', 'Std [s]', 'MaxAlloc. CUDA mem [MB]']
+        self.table_fields = ['Radius [Å]', 'Num. atoms', 'Mean [s]', 'Std [s]', 'MaxAlloc. CUDA Memory (Structure) [MB]', 'MaxAlloc. CUDA Memory (Calc) [MB]']
         self.pt = PrettyTable(self.table_fields)
         self.pt.align = 'r'
         self.pt.padding_width = 1
         self.pt.title = 'Benchmark Generator / ' + self.device + ' / Batch Size: ' + str(self.batch_size)
-        self.data = [[str(float(r)), str(int(n)), f'{m:1.5f}', f'{s:1.5f}', f'{c:1.5f}'] for r,n,m,s,c in zip(self.radii, list(num_atoms), list(means), list(stds), list(cuda_mem))]
+        self.data = [[str(float(r)), str(int(n)), f'{m:1.5f}', f'{s:1.5f}', f'{cs:1.5f}', f'{cc:1.5f}'] for r,n,m,s,cs,cc in zip(self.radii, list(num_atoms), list(means), list(stds), list(cuda_mem_structure), list(cuda_mem_calculations))]
         for d in self.data:
             self.pt.add_row(d)
 
@@ -85,7 +87,7 @@ class Statistics:
         """
         Return a detailed string representation of the Statistics object.
         """
-        return f'Statistics (\n\tname = {self.name},\n\tradii = {self.radii},\n\tnum_atoms = {self.num_atoms},\n\tmeans = {self.means},\n\tstds = {self.stds},\n\tcuda_mem = {self.cuda_mem},\n\tbatch_size = {self.batch_size},\n\tdevice = {self.device}\n)'
+        return f'Statistics (\n\tname = {self.name},\n\tradii = {self.radii},\n\tnum_atoms = {self.num_atoms},\n\tmeans = {self.means},\n\tstds = {self.stds},\n\tcuda_mem_structure = {self.cuda_mem_structure},\n\tcuda_mem_calculations = {self.cuda_mem_calculations},\n\tbatch_size = {self.batch_size},\n\tdevice = {self.device}\n)'
 
 class DebyeBenchmarker:
     """
@@ -100,6 +102,7 @@ class DebyeBenchmarker:
     """
     def __init__(
         self,
+        function: str = 'gr',
         radii: Union[List, np.ndarray, torch.Tensor] = [5],
         show_progress_bar: bool = True,
         custom_cif: str = None,
@@ -118,11 +121,21 @@ class DebyeBenchmarker:
         self.set_radii(list(radii))
         self.cif = pkg_resources.resource_filename(__name__, 'benchmark_structure.cif')
         self.custom_cif = custom_cif
-
-        self.show_progress_bar = show_progress_bar
+        
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
             self.debye_calc = DebyeCalculator(**kwargs)
+            
+        if function == 'gr':
[…]
