+++ debyecalculator/utility/benchmark.py
@@ -33,46 +33,50 @@ class Statistics:
     """
     def __init__(
         self,
-        means: List[float],
-        stds: List[float],
         name: str,
+        function_name: str,
+        device: str,
+        batch_size: int,
         radii: List[float],
         num_atoms: List[int],
+        means: List[float],
+        stds: List[float],
         cuda_mem_structure: List[float],
         cuda_mem_calculations: List[float],
-        device: str,
-        batch_size: int
     ) -> None:
         """
         Initialize Statistics with benchmarking results.
 
         Parameters:
-        - means (List[float]): List of mean values.
-        - stds (List[float]): List of standard deviation values.
         - name (str): Name of the statistics.
-        - radii (List[float]): List of radii.
-        - num_atoms (List[int]): List of number of atoms.
-        - cuda_mem (List[float]): List of CUDA memory values.
+        - function_name (str) Name of function to be benchmarked
         - device (str): Device used for benchmarking.
         - batch_size (int): Batch size used for benchmarking.
+        - means (List[float]): List of mean values.
+        - stds (List[float]): List of standard deviation values.
+        - radii (List[float]): List of radii.
+        - num_atoms (List[int]): List of number of atoms.
+        - cuda_mem_structure (List[float]): List of CUDA memory values for generating structures.
+        - cuda_mem_calculations (List[float]): List of CUDA memory values for calculating the function.
         """
 
+        self.name = name
+        self.function_name = function_name
+        self.device = device
+        self.batch_size = batch_size
         self.means = means
         self.stds = stds
-        self.name = name
         self.radii = radii
         self.num_atoms = num_atoms
         self.cuda_mem_structure = cuda_mem_structure
         self.cuda_mem_calculations = cuda_mem_calculations
-        self.device = device
-        self.batch_size = batch_size
         
         # Create table
-        self.table_fields = ['Radius [Å]', 'Num. atoms', 'Mean [s]', 'Std [s]', 'MaxAlloc. CUDA Memory (Structure) [MB]', 'MaxAlloc. CUDA Memory (Calc) [MB]']
+        self.table_fields = ['Radius [Å]', 'Num. atoms', 'Mean [s]', 'Std [s]', 'MaxAllocCUDAMem (Gen.) [MB]', 'MaxAllocCUDAMem (Calc.) [MB]']
         self.pt = PrettyTable(self.table_fields)
-        self.pt.align = 'r'
+        self.pt.align = 'c'
         self.pt.padding_width = 1
-        self.pt.title = 'Benchmark Generator / ' + self.device + ' / Batch Size: ' + str(self.batch_size)
+        self.pt.title = 'Benchmark / ' + self.function_name + ' / ' + self.device.capitalize() + ' / Batch Size: ' + str(self.batch_size)
         self.data = [[str(float(r)), str(int(n)), f'{m:1.5f}', f'{s:1.5f}', f'{cs:1.5f}', f'{cc:1.5f}'] for r,n,m,s,cs,cc in zip(self.radii, list(num_atoms), list(means), list(stds), list(cuda_mem_structure), list(cuda_mem_calculations))]
         for d in self.data:
             self.pt.add_row(d)
@@ -87,7 +91,7 @@ class Statistics:
         """
         Return a detailed string representation of the Statistics object.
         """
[…]
