### 2024-01-21 4bf9f6b
simplified generation script for benchmarking purposes, needs refactoring
