+++ debyecalculator/utility/generate.py
@@ -102,6 +102,7 @@ def generate_nanoparticles(
     _lightweight_mode: bool = False,
     _return_ase: bool = False,
     _reverse_order: bool = True,
+    _benchmarking: bool = False,
 ) -> NanoParticleType:
     """
     Generate spherical nanoparticles from a given CIF and radii.
@@ -118,6 +119,8 @@ def generate_nanoparticles(
         _override_device (bool): Ignore object device and run on CPU.
         _lightweight_mode (bool): Whether to use lightweight mode. Defaults to False.
         _return_ase (bool): Whether to return ASE objects. Defaults to False.
+        _reverse_order (bool): Whether to generate particles in reverse radii order
+        _benchmarking (bool): Stripped down version for benchmarking
 
     Returns:
         NanoParticleType: List of nanoparticle tuples or ASE objects.
@@ -194,6 +197,27 @@ def generate_nanoparticles(
 
     # Convert positions to torch and send to device
     positions = torch.from_numpy(cell.get_positions()).to(dtype = torch.float32, device = device)
+    
+    # Benchmarking
+    if _benchmarking:
+        nanoparticle_tuple_list = []
+        for r in sorted(radii, reverse=_reverse_order):
+            cell_norms = torch.norm(positions, p=2, dim=-1).cpu()
+            np_cell = cell[cell_norms <= r]
+            elements = np_cell.get_chemical_symbols()
+            try:
+                occupancy = np_cell.info['occupancy']
+            except:
+                occupancy = torch.ones((np_cell.get_global_number_of_atoms()), dtype=torch.float32)
+            nanoparticle_tuple_list.append(
+                NanoParticle(
+                    elements = elements,
+                    size = len(elements),
+                    occupancy = occupancy,
+                    xyz = torch.from_numpy(np_cell.get_positions()).to(device=device)
+                )
+            )
+        return nanoparticle_tuple_list
 
     # Find atomic radii
     atomic_radii = torch.tensor(np.array([
