### 2024-01-21 edab01d
added TITANRTX reference and plotting
