### 2024-01-21 ec25446
added doc strings for plotting functions
