### 2024-01-21 FrederikLizakJohansen/DebyeCalculator#19 @FrederikLizakJohansen
Thank you for your suggestion. We find this a very relevant feature and so we wanted to incorporate it into DebyeCalculator. You can see the additions we've made in merge-commit 718be48.

**To summarize;** We have added a benchmarker class (`DebyeBenchmarker`) under the utility folder including some additional functions to save, load and plot the benchmarks. The class comes with (so far) two reference benchmarks, including one run using DiffPy-CMI on CPU and one run on a CUDA node with a TITAN RTX with a batch size of 10k. The class has its own default structure (CIF) file that it uses for benchmarking, but optionally a custom cif can be used. The benchmarking now also tries to keep track of the allocated memory on the CUDA device. Additionally, we have added a benchmarking notebook [Benchmarking.ipynb](https://github.com/FrederikLizakJohansen/DebyeCalculator/blob/main/Benchmark.ipynb) that guide the user through using the benchmarking class and comparing with the reference.

We trust that these additions meet your expectations. We will close this issue here. Feel free to reopen the issue if you would like to see any further additions made to this and we will consider those in due time.
