### 2024-01-21 FrederikLizakJohansen/DebyeCalculator#29 @FrederikLizakJohansen
Thank you for your feedback and suggestion.
The data that is being used to generate the benchmarking figure is included in the data folder that comes with DebyeCalculator and all structures are generated using the `generate_nanoparticles` function that is available to the user. We understand the importance of transparency, so we have added a link to the exact data-file that is being used in the caption of the figure.

There was a suggestion in #19 to implement a benchmarking script, which we ended up doing. You can read more about that in that issue. We believe these additions will significantly improve the utility and clarity of the benchmarking process and make it clear to the user how much they can gain when moving to the GPU.

I will close this issue for now, but feel free to reopen it if you have any other concerns regarding the data.
