### 2024-01-21 2855a56
fixed generation assertion
