### 2024-01-22 a3b7a9d
added benchmarking to readthedocs
