+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 sphinx
 sphinx_rtd_theme
+yaml
