+++ docs/requirements.txt
@@ -1,3 +1,3 @@
 sphinx
 sphinx_rtd_theme
-yaml
+PyYAML
