+++ docs/requirements.txt
@@ -10,4 +10,3 @@ setuptools
 tqdm
 ase
 debyecalculator
-torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
