+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
-  #os: ubuntu-22.04
+  os: ubuntu-22.04
   tools:
     python: "3.7"
     # You can also specify other tool versions:
