+++ docs/conf.py
@@ -11,6 +11,8 @@ import sphinx_rtd_theme
 html_theme = "sphinx_rtd_theme"
 html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
+import warnings
+warnings.filterwarnings("ignore", category=UserWarning, module="debyecalculator")
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
