+++ .readthedocs.yaml
@@ -18,7 +18,8 @@ build:
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/conf.py
-
+  fail_on_warning: false
+  
 # Optionally build your docs in additional formats such as PDF and ePub
 # formats:
 #    - pdf
