+++ .readthedocs.yaml
@@ -6,10 +6,10 @@
 version: 2
 
 # Set the OS, Python version and other tools you might need
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.7"
+#build:
+#  os: ubuntu-22.04
+#  tools:
+#    python: "3.7"
     # You can also specify other tool versions:
     # nodejs: "19"
     # rust: "1.64"
@@ -17,6 +17,7 @@ build:
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
+  builder: html
   configuration: docs/conf.py
   fail_on_warning: false
   
@@ -28,6 +29,6 @@ sphinx:
 # Optional but recommended, declare the Python requirements required
 # to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-python:
-   install:
-   - requirements: docs/requirements.txt
+#python:
+#   install:
+#   - requirements: docs/requirements.txt
