+++ .readthedocs.yaml
@@ -6,10 +6,10 @@
 version: 2
 
 # Set the OS, Python version and other tools you might need
-#build:
-#  os: ubuntu-22.04
-#  tools:
-#    python: "3.7"
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.7"
     # You can also specify other tool versions:
     # nodejs: "19"
     # rust: "1.64"
