### 2024-01-22 afcfc7a
Update .readthedocs.yaml
