+++ .readthedocs.yaml
@@ -17,7 +17,7 @@ build:
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
-  #builder: html
+  builder: html
   configuration: docs/conf.py
   fail_on_warning: false
   
