+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 sphinx
-
+sphinx_rtd_theme
 PyYAML
 prettytable
 numpy
