+++ .readthedocs.yaml
@@ -28,6 +28,6 @@ sphinx:
 # Optional but recommended, declare the Python requirements required
 # to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-#python:
-#   install:
-#   - requirements: docs/requirements.txt
+python:
+   install:
+   - requirements: docs/requirements.txt
