+++ docs/conf.py
@@ -2,6 +2,8 @@
 #
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
+autodoc_mock_imports = ["torch", "debyecalculator"]
+
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
