+++ debyecalculator/debye_calculator.py
@@ -29,6 +29,11 @@ except ModuleNotFoundError:
         "https://github.com/FrederikLizakJohansen/DebyeCalculator"
     )
 '''
+try:
+    import torch
+except ImportError:
+    torch = None
+
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -177,6 +182,14 @@ class DebyeCalculator:
         
         # Lightweight mode
         self._lightweight_mode = _lightweight_mode
+
+        if torch is None:
+            raise RuntimeError("\n\nDebyeCalculator requires PyTorch, which is not installed. "
+                                "Please install PyTorch before using DebyeCalculator. "
+                                "Follow the instructions on the official PyTorch website: "
+                                "https://pytorch.org/get-started/locally/. "
+                                "For more information about DebyeCalculator, visit the GitHub repository: "
+                                "https://github.com/FrederikLizakJohansen/DebyeCalculator")
 
     def __repr__(
         self,
