+++ debyecalculator/debye_calculator.py
@@ -29,10 +29,6 @@ except ModuleNotFoundError:
         "https://github.com/FrederikLizakJohansen/DebyeCalculator"
     )
 '''
-try:
-    import torch
-except ImportError:
-    torch = None
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -183,13 +179,6 @@ class DebyeCalculator:
         # Lightweight mode
         self._lightweight_mode = _lightweight_mode
 
-        if torch is None:
-            raise RuntimeError("\n\nDebyeCalculator requires PyTorch, which is not installed. "
-                                "Please install PyTorch before using DebyeCalculator. "
-                                "Follow the instructions on the official PyTorch website: "
-                                "https://pytorch.org/get-started/locally/. "
-                                "For more information about DebyeCalculator, visit the GitHub repository: "
-                                "https://github.com/FrederikLizakJohansen/DebyeCalculator")
 
     def __repr__(
         self,
