### 2024-01-22 dc11759
fixed doc strings
