+++ debyecalculator/utility/benchmark.py
@@ -20,16 +20,6 @@ from collections import namedtuple
 class Statistics:
     """
     A class to store and represent benchmark statistics.
-
-    Attributes:
-    - means (List[float]): List of mean values.
-    - stds (List[float]): List of standard deviation values.
-    - name (str): Name of the statistics.
-    - radii (List[float]): List of radii.
-    - num_atoms (List[int]): List of number of atoms.
-    - cuda_mem (List[float]): List of CUDA memory values.
-    - device (str): Device used for benchmarking.
-    - batch_size (int): Batch size used for benchmarking.
     """
     def __init__(
         self,
@@ -48,16 +38,16 @@ class Statistics:
         Initialize Statistics with benchmarking results.
 
         Parameters:
-        - name (str): Name of the statistics.
-        - function_name (str) Name of function to be benchmarked
-        - device (str): Device used for benchmarking.
-        - batch_size (int): Batch size used for benchmarking.
-        - means (List[float]): List of mean values.
-        - stds (List[float]): List of standard deviation values.
-        - radii (List[float]): List of radii.
-        - num_atoms (List[int]): List of number of atoms.
-        - cuda_mem_structure (List[float]): List of CUDA memory values for generating structures.
-        - cuda_mem_calculations (List[float]): List of CUDA memory values for calculating the function.
+            name (str): Name of the statistics.
+            function_name (str) Name of function to be benchmarked
+            device (str): Device used for benchmarking.
+            batch_size (int): Batch size used for benchmarking.
+            radii (List[float]): List of radii.
+            num_atoms (List[int]): List of number of atoms.
+            means (List[float]): List of mean values.
+            stds (List[float]): List of standard deviation values.
+            cuda_mem_structure (List[float]): List of CUDA memory values for generating structures.
+            cuda_mem_calculations (List[float]): List of CUDA memory values for calculating the function.
         """
 
         self.name = name
@@ -87,27 +77,23 @@ class Statistics:
 
     def __str__(self) -> str:
         """
-        Return a PrettyTable Statistics table.
+        Returns:
+            A PrettyTable Statistics table.
         """
         return str(self.pt)
 
     def __repr__(self) -> str:
         """
-        Return a detailed string representation of the Statistics object.
+        Returns:
+            A detailed string representation of the Statistics object.
         """
         return f'Statistics (\n\tname = {self.name},\n\tfunction_name = {self.function_name},\n\tdevice = {self.device},\n\tbatch_size = {self.batch_size},\n\tradii = {self.radii},\n\tnum_atoms = {self.num_atoms},\n\tmeans = {self.means},\n\tstds = {self.stds},\n\tcuda_mem_structure = {self.cuda_mem_structure},\n\tcuda_mem_calculations = {self.cuda_mem_calculations},\n)'
 
 class DebyeBenchmarker:
     """
     A class for benchmarking Debye calculations.
-
-    Attributes:
-    - radii (List[float]): List of radii for benchmarking.
-    - cif (str): Path to reference CIF file used for benchmarking.
-    - custom_cif (str): Custom CIF file path (if provided).
-    - show_progress_bar (bool): Flag to control progress bar display.
-    - debye_calc (DebyeCalculator): Debye calculator instance.
     """
+
     def __init__(
         self,
         function: str = 'gr',
@@ -120,10 +106,12 @@ class DebyeBenchmarker:
         Initialize DebyeBenchmarker.
 
         Parameters:
-        - radii (Union[List, np.ndarray, torch.Tensor]): List of radii for benchmarking.
-        - show_progress_bar (bool): Flag to control progress bar display.
-        - custom_cif (str): Custom CIF file path (if provided).
[…]
