### 2024-01-22 FrederikLizakJohansen/DebyeCalculator#13 @FrederikLizakJohansen
We have now tagged an official release on Github for our version 1.0.7: [Release Notes](https://github.com/FrederikLizakJohansen/DebyeCalculator/releases/tag/v1.0.7).

I will close this issue for now.
