### 2024-01-22 #6024 @AndySAnker
We have today finalised our answer to the first review round with the release of `DebyeCalculator` v.1.0.7.!

Our gratitude goes to @KedoKudo @marcocamma @elena-pascal for their work and suggestions for revisions, which have substantially enhanced the software.

We apologise for any delays in our response, which were required to make several of the updates to DebeyeCalaculator that are outlined below.

For a detailed log of the modifications, kindly refer to [issues](https://github.com/FrederikLizakJohansen/DebyeCalculator/issues?q=is%3Aissue+is%3Aclosed) on the GitHub page.
Here is a summary of the improvements made to`DebyeCalculator`:
•	A benchmarking clas has been released to include various scenarios including atom count, GPU usage, and batch sizes.
•	We have conducted the benchmarking across 2 different hardware setups. Results are directly reproducible from the script to the associated figure in the paper.
•	Clear version tagging and release notes have been instituted.
•	We have refined our naming conventions for greater clarity.
•	PyTorch installation has been shifted to a user-led process, paired with advanced error messaging for troubleshooting installation issues.
•	Testing instructions have been implemented.
•	Input structure flexibility has been enhanced to support tuples, facilitating the calculation of partial Gr.
•	`DebyeCalculator` is now compatible with Python3.7 and Python3.8.
•	Refined installation guidelines are now available.
•	Documentation is now on Read the Docs, available at (https://debyecalculator.readthedocs.io/en/latest/).

There are few feature enhancements (some) arising from the review process that we would like to add to `DebyeCalculator` in future releases, such as:
•	The possibility of extracting differently shaped discrete structures from a CIF file and calculating the scattering pattern.
•	Enable the calculation of very large structures (many atoms).
•	Make it more intuitive to upload data to the Google Colab version of `DebyeCalculator`.
•	Enable `DebyeCalculator` to be run on the GPU on a MacBook.

Currently, we are prioritising the release of robust and user-friendly software, with further advancements to be integrated progressively.

With these substantial modifications, that have addressed majority of the concerns from the reviewers, and we hope that you find our DebeyeCalculator paper suitable for publication in the Journal of Open Source Software.
