### 2024-01-26 #6024 @elena-pascal
I went through my checklist and everything looks good to me. I checked the performance claims and they seem reasonable. Thank you for the hard work put into documenting the package, it makes it significantly more user friendly. 👍
