### 2024-02-14 c9e1f82
Update paper.md
Updated affiliation on Andy and date.
