### 2024-02-14 5a4f741
Update paper.md
Updated Raghavs affiliation
