### 2024-02-14 #6024 @AndySAnker
Thank you @KedoKudo @marcocamma @elena-pascal for the help and feedback with the paper.

@phibeck Thank you for editing this. We have:
- Double checked authors and affiliations (including ORCIDs)
- Made a release of the software with the latest changes from the review. The version is **v1.0.7**.
- Archived the release on Zenodo. The DOI is: **10.5281/zenodo.10659528**
- Made sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- Made sure that the license listed for the archive is the same as the software license.

Let us know, if there is anything you need from us.
