### 2024-02-19 #6024 @phibeck
@AndySAnker here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.
- line 12 and all other citations: please add a whitespace before references
- line 36: I would use backticks for `` ` ``DebyeCalculator`` ` ``, to produce `DebyeCalculator`
- line 37: similarly for xyz-files or CIF-files I would opt for `.xyz` files or `.CIF` files
- line 45f: please insert the links to GitHub and PyPI as hyperlinks
- line 50: citation before comma
- line 52: perhaps link to Fig. 3 directly in Supporting Information or have a section link to SI
- Figure 1 caption Fig 1, last line: please use a hyperlink
- line 78: I would prefer: "Moreover, it can be installed conveniently via `pip`..."
- line 78: please fix hyperlink

References:
- line 92: Debye; missing DOI
- line 94: Debye; missing DOI
- line 96: Röntgenstrahlen; Annalen der Physik; missing DOI
- line 97: Bragg
- line 99: missing DOI
- line 103: please correct the rendering of the material $V_{0.985}Al_{0.015}O_2$
- line 110f: missing DOI
- line 112f: missing DOI
- line 114f: BCL::SAXS:; Debye; X-ray
- line 117: Debye; missing DOI
- line 120: missing DOI
