### 2024-02-19 1f77bcd
Update paper.md
Updated based on Editor comments in during review process:
https://github.com/openjournals/joss-reviews/issues/6024#issuecomment-1953134439
