+++ paper/paper.md
@@ -84,7 +84,7 @@ CLASS `DebyeCalculator`:
       - Return G(r) either on GPU or CPU            
 ```
 
-In order to benchmark our implementation, we compare simulated scattering patterns from `DebyeCalculator` against DiffPy-CMI [@juhas2015complex], which is a widely recognised software for scattering pattern computations. Here, our implementation obtains the same scattering patterns as DiffPy-CMI (\autoref{fig:figure_3}), while being faster on CPU for structures up to ~20,000 atoms (\autoref{fig:figure_1}). Both calculations are run on a x86-64 CPU with 64GB of memory and a batch size of 10,000.
+In order to benchmark our implementation, we compare simulated scattering patterns from `DebyeCalculator` against DiffPy-CMI [@juhas2015complex], which is a widely recognised software for scattering pattern computations. Here, our implementation obtains the same scattering patterns as DiffPy-CMI (\autoref{fig:figure_S1}), while being faster on CPU for structures up to ~20,000 atoms (\autoref{fig:figure_1}). Both calculations are run on a x86-64 CPU with 64GB of memory and a batch size of 10,000.
 Running the calculations on the GPU provides another notable boost in speed (\autoref{fig:figure_1}). This improvement primarily stems from the distribution of the double sum calculations across a more extensive set of cores than is feasible on the CPU. With smaller atomic structures, an overhead associated with initiating GPU calculations results in the NVIDIA RTX A3000 Laptop GPU computations being slower than DiffPy-CMI and our CPU implementation. Once the atomic structure size exceeds ~14 Å in diameter (~300 atoms), we observe a ~5 times speed-up using an NVIDIA RTX A3000 Laptop GPU with 6GB of memory and a batch size of 10,000. 
 The choice of GPU hardware has a substantial influence on this speed advantage. As demonstrated in \autoref{fig:figure_1}, using an NVIDIA Titan RTX GPU, which offers 24GB of memory, the speed benefits become even more evident. The NVIDIA Titan RTX GPU delivers a performance that is ~10 times faster, seemingly across all structure sizes, underlining the significant role of the hardware. With the advancements of GPUs like NVIDIA's Grace Hopper Superchip [@NVIDIA], which boasts 576GB of fast-access to memory, there is potential for `DebyeCalculator` to achieve even greater speeds in the future.
 
