### 2024-02-19 312ffaf
Update paper.bib
Added DOI's
