+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={589--590},
   year={2016},
   publisher={International Union of Crystallography}
+  doi={10.1107/S2053273316015680}
 }
 
 @article{debye:1915,
@@ -18,6 +19,7 @@
   pages={809--823},
   year={1915},
   publisher={Wiley Online Library}
+  doi={10.1002/andp.19153510606}
 }
 
 @article{schaller1997moore,
@@ -29,6 +31,7 @@
   pages={52--59},
   year={1997},
   publisher={IEEE}
+  doi={10.1109/6.591665}
 }
 
 @article{garland2008parallel,
@@ -40,6 +43,7 @@
   pages={13--27},
   year={2008},
   publisher={IEEE}
+  doi={10.1109/MM.2008.57}
 }
 
 @book{egami2003underneath,
@@ -71,6 +75,7 @@
   pages={1543--1547},
   year={2010},
   publisher={International Union of Crystallography}
+  doi={10.1107/S0021889810041889}
 }
 
 @article{cervellino2015debussy,
@@ -82,6 +87,7 @@
   pages={2026--2032},
   year={2015},
   publisher={International Union of Crystallography}
+  doi={10.1107/S1600576715020488}
 }
 
 @article{coelho2018topas,
@@ -93,6 +99,7 @@
   pages={210--218},
   year={2018},
   publisher={International Union of Crystallography}
+  doi={10.1107/S1600576718000183}
 }
 
 @article{proffen1999discus,
@@ -104,6 +111,7 @@
   pages={838--839},
   year={1999},
   publisher={International Union of Crystallography}
+  doi={10.1107/S0021889899004860}
 }
 
 @article{proffen1997discus,
@@ -115,6 +123,7 @@
   pages={171--175},
   year={1997},
   publisher={International Union of Crystallography}
+  doi={10.1107/S002188989600934X}
 }
 
 @article{putnam2015bcl,
