+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={6},
   pages={589--590},
   year={2016},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S2053273316015680}
 }
 
@@ -18,7 +18,7 @@
   number={6},
   pages={809--823},
   year={1915},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/andp.19153510606}
 }
 
@@ -30,7 +30,7 @@
   number={6},
   pages={52--59},
   year={1997},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/6.591665}
 }
 
@@ -42,7 +42,7 @@
   number={4},
   pages={13--27},
   year={2008},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/MM.2008.57}
 }
 
@@ -74,7 +74,7 @@
   number={6},
   pages={1543--1547},
   year={2010},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S0021889810041889}
 }
 
@@ -86,7 +86,7 @@
   number={6},
   pages={2026--2032},
   year={2015},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S1600576715020488}
 }
 
@@ -98,7 +98,7 @@
   number={1},
   pages={210--218},
   year={2018},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S1600576718000183}
 }
 
@@ -110,7 +110,7 @@
   number={4},
   pages={838--839},
   year={1999},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S0021889899004860}
 }
 
@@ -122,7 +122,7 @@
   number={2},
   pages={171--175},
   year={1997},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
   doi={10.1107/S002188989600934X}
 }
 
