+++ paper/paper.bib
@@ -127,7 +127,7 @@
 }
 
 @article{putnam2015bcl,
-  title={BCL:: SAXS: GPU accelerated Debye method for computation of small angle X-ray scattering profiles},
+  title={BCL::SAXS: GPU accelerated Debye method for computation of small angle X-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
   journal={Proteins: Struct., Funct., Genet.},
   volume={83},
@@ -152,7 +152,7 @@ abstract = {New analytical X-ray scattering-factor representations valid for the
 }
 
 @article{GHEDIRA1977423,
-title = {Structural aspects of the metal-insulator transitions in V~0.985~Al~0.015O2~},
+title = {Structural aspects of the metal-insulator transitions in V~0.985~Al~0.015~O~2~},
 journal = {J. Solid State Chem.},
 volume = {22},
 number = {4},
