+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{scardi:2016,
-  title={Celebrating 100 years of the Debye scattering equation},
+  title={Celebrating 100 years of the {D}ebye scattering equation},
   author={Scardi, Paolo and Billinge, Simon JL and Neder, Reinhard and Cervellino, Antonio},
   journal={Acta Crystallogr. A},
   volume={72},
