+++ paper/paper.bib
@@ -13,7 +13,7 @@
 @article{debye:1915,
   title={Zerstreuung von r{\"o}ntgenstrahlen},
   author={Debye, Peter},
-  journal={Annalen der Physik},
+  journal={Annalen {d}er Physik},
   volume={351},
   number={6},
   pages={809--823},
@@ -47,7 +47,7 @@
 }
 
 @book{egami2003underneath,
-  title={Underneath the Bragg peaks: structural analysis of complex materials},
+  title={Underneath the {B}ragg peaks: structural analysis of complex materials},
   author={Egami, Takeshi and Billinge, Simon JL},
   year={2003},
   publisher={Elsevier},
@@ -127,7 +127,7 @@
 }
 
 @article{putnam2015bcl,
-  title={BCL::SAXS: GPU accelerated Debye method for computation of small angle X-ray scattering profiles},
+  title={BCL::SAXS: GPU accelerated {D}ebye method for computation of small angle X-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
   journal={Proteins: Struct., Funct., Genet.},
   volume={83},
@@ -167,4 +167,4 @@ author = {M. Ghedira and H. Vincent and M. Marezio and J.C. Launay}
 
 @misc{CIF, url={https://github.com/FrederikLizakJohansen/DebyeCalculatorGPU/blob/main/example_data/AntiFluorite_Co2O.cif}, journal={GitHub}, author={Ulrik Friis-Jensen}, year={2023}}
 
-@misc{debyer, title={Wojdyr/debyer: Debye’s scattering equation and other analysis of atomistic models.}, url={https://github.com/wojdyr/debyer}, journal={GitHub}, author={Wojdyr}, year={2023}} 
+@misc{debyer, title={Wojdyr/debyer: {D}ebye’s scattering equation and other analysis of atomistic models.}, url={https://github.com/wojdyr/debyer}, journal={GitHub}, author={Wojdyr}, year={2023}} 
