+++ paper/paper.bib
@@ -67,7 +67,7 @@
 }
 
 @article{cervellino2010debussy,
-  title={DEBUSSY: a Debye user system for nanocrystalline materials},
+  title={DEBUSSY: a {D}ebye user system for nanocrystalline materials},
   author={Cervellino, Antonio and Giannini, Cinzia and Guagliardi, Antonietta},
   journal={J. Appl. Crystallogr.},
   volume={43},
@@ -79,7 +79,7 @@
 }
 
 @article{cervellino2015debussy,
-  title={DEBUSSY 2.0: the new release of a Debye user system for nanocrystalline and/or disordered materials},
+  title={DEBUSSY 2.0: the new release of a {D}ebye user system for nanocrystalline and/or disordered materials},
   author={Cervellino, Antonio and Frison, Ruggero and Bertolotti, Federica and Guagliardi, Antonietta},
   journal={J. Appl. Crystallogr.},
   volume={48},
@@ -127,7 +127,7 @@
 }
 
 @article{putnam2015bcl,
-  title={BCL::SAXS: GPU accelerated {D}ebye method for computation of small angle X-ray scattering profiles},
+  title={BCL::SAXS: GPU accelerated {D}ebye method for computation of small angle {X}-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
   journal={Proteins: Struct., Funct., Genet.},
   volume={83},
@@ -152,7 +152,7 @@ abstract = {New analytical X-ray scattering-factor representations valid for the
 }
 
 @article{GHEDIRA1977423,
-title = {Structural aspects of the metal-insulator transitions in V~0.985~Al~0.015~O~2~},
+title = {Structural aspects of the metal-insulator transitions in {V~0.985~Al~0.015~O~2~}},
 journal = {J. Solid State Chem.},
 volume = {22},
 number = {4},
