+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{debye:1915,
-  title={Zerstreuung von r{\"o}ntgenstrahlen},
+  title={Zerstreuung von {R}{\"o}ntgenstrahlen},
   author={Debye, Peter},
   journal={Annalen {d}er Physik},
   volume={351},
