+++ paper/paper.bib
@@ -152,7 +152,7 @@ abstract = {New analytical X-ray scattering-factor representations valid for the
 }
 
 @article{GHEDIRA1977423,
-title = {Structural aspects of the metal-insulator transitions in {V~0.985~Al~0.015~O~2~}},
+title = {Structural aspects of the metal-insulator transitions in {V_{0.985} Al_{0.015} O_{2}}},
 journal = {J. Solid State Chem.},
 volume = {22},
 number = {4},
