+++ paper/paper.bib
@@ -127,7 +127,7 @@
 }
 
 @article{putnam2015bcl,
-  title={BCL::SAXS: GPU accelerated {D}ebye method for computation of small angle {X}-ray scattering profiles},
+  title={BCL::SAXS: GPU accelerated {D}ebye method for computation of small-angle {X}-ray scattering profiles},
   author={Putnam, Daniel K and Weiner, Brian E and Woetzel, Nils and Lowe Jr, Edward W and Meiler, Jens},
   journal={Proteins: Struct., Funct., Genet.},
   volume={83},
