### 2024-02-20 4ae06f1
Updated based on Editor comments in during review process: openjournals/joss-reviews#6024
