### 2024-02-20 #6024 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
