### 2024-02-20 #6024 @kyleniemeyer
Hi @AndySAnker, I'm doing some final checks before publishing.

My only comment is that the reference to the NVIDIA article seems a bit off - there's no title, and normally for a web page you would add a note about when it was accessed (like `note = {Accessed: 2024-02-20}`.
