+++ paper/paper.bib
@@ -163,7 +163,7 @@ doi = {10.1016/0022-4596(77)90020-2},
 author = {M. Ghedira and H. Vincent and M. Marezio and J.C. Launay}
 }
 
-@misc{NVIDIA, title={NVIDIA Grace Hopper Superchip Data Sheet}, url={https://resources.nvidia.com/en-us-grace-cpu/grace-hopper-superchip - Accessed: 2024-02-20}, journal={NVIDIA}, author={NVIDIA}, year={2024}} 
+@misc{NVIDIA, title={NVIDIA Grace Hopper Superchip Data Sheet - Accessed: 2024-02-20}, url={https://resources.nvidia.com/en-us-grace-cpu/grace-hopper-superchip}, journal={NVIDIA}, author={NVIDIA}, year={2024}} 
 
 @misc{CIF, url={https://github.com/FrederikLizakJohansen/DebyeCalculatorGPU/blob/main/example_data/AntiFluorite_Co2O.cif}, journal={GitHub}, author={Ulrik Friis-Jensen}, year={2023}}
 
