### 2024-02-20 #6024 @AndySAnker
Hi @kyleniemeyer, that is a great point - thanks!

I have now updated the paper to state:
"_With the advancements of GPUs like NVIDIA's Grace Hopper Superchip [@NVIDIA], which boasts 624GB of fast-access to memory_" Instead of "_576GB_" which was the fast-access memory of the chip when we submitted the paper.

I have also added a title and a note to the reference, which now looks like this:

If we add the note as suggested, it does not appear in the reference and readers would have to open the .bib file to see the note. Therefore, I have added this to the title.
